% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{add_dashed_lines}
\alias{add_dashed_lines}
\title{Add dashed lines for treatment timepoints to a plot}
\usage{
add_dashed_lines(p, treatment_timepoints, treatment_labels, y_pos)
}
\arguments{
\item{p}{A ggplot object. The plot to which dashed lines and labels 
will be added.}

\item{treatment_timepoints}{A numeric vector of timepoints where 
dashed lines should be drawn.}

\item{treatment_labels}{A character vector of labels corresponding 
to each treatment timepoint. These labels are used for coloring 
the lines, but the x-axis coordinates are displayed as the labels.}

\item{y_pos}{A numeric value specifying the y-axis position where 
the text labels should be placed.}
}
\value{
A ggplot object with added dashed lines and labels.
}
\description{
This internal function adds dashed vertical lines at specified 
treatment timepoints to a plot, along with text labels that 
display the corresponding x-axis values.
}
