% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{add_gene_column}
\alias{add_gene_column}
\title{Add Gene Column}
\usage{
add_gene_column(df, genes)
}
\arguments{
\item{df}{A tibble containing a `feature_index` column with numeric values.}

\item{genes}{A character vector of gene names.}
}
\value{
A tibble with an added `Gene` column placed after the `feature_index`
 column.
 
 @importFrom dplyr mutate
}
\description{
This function takes a tibble and a character vector of genes, and adds a 
column named `Gene` to the tibble. The `Gene` column is created based on the 
values in the `feature_index` column, and is placed immediately after it.
}
