% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{add_plot_to_html}
\alias{add_plot_to_html}
\title{Add Plot to HTML Content}
\usage{
add_plot_to_html(html_content, plot_element, plots_size, section_index)
}
\arguments{
\item{html_content}{The current HTML content as a character string.}

\item{plot_element}{The plot element to be converted to base64.}

\item{plots_size}{An integer specifying the height of the plot.}

\item{section_index}{An integer specifying the section index.}
}
\value{
The updated HTML content as a character string.
}
\description{
This function converts a plot to a base64 image and adds it to the
HTML content.
}
