% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{ask_user}
\alias{ask_user}
\title{Prompt the user with a yes/no question}
\usage{
ask_user(question)
}
\arguments{
\item{question}{A string of the question to ask the user.}
}
\value{
None.
}
\description{
This function prompts the user with a yes/no question. If the user answers
"yes" (case insensitive), the code proceeds. If the user answers "no" or
anything else, the code stops.
}
