% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{bind_data_with_annotation}
\alias{bind_data_with_annotation}
\title{Bind Data with Annotation}
\usage{
bind_data_with_annotation(data, annotation = NULL)
}
\arguments{
\item{data}{A matrix containing the numeric data.}

\item{annotation}{A dataframe containing the annotation information.}
}
\value{
A dataframe with `data` and `annotation` combined, and the row names
 of `data`
as the first column named `feature_names`.
}
\description{
This function converts a matrix to a dataframe, adds row names as the first 
column,
and binds it with annotation data.
}
