% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{build_hyperparams_screen_report}
\alias{build_hyperparams_screen_report}
\title{Build Hyperparameters Screening Report}
\usage{
build_hyperparams_screen_report(
  header_section,
  plots,
  plots_sizes,
  output_file_path
)
}
\arguments{
\item{header_section}{A character string containing the HTML header section.}

\item{plots}{A list of ggplot2 plot objects.}

\item{plots_sizes}{A list of integers specifying the number of rows for each 
plot.}

\item{output_file_path}{A character string specifying the path to save the 
HTML report.}
}
\value{
No return value, called for side effects.
}
\description{
Constructs an HTML report for hyperparameter screening by embedding plots 
and their respective sizes into the provided header section.
}
\seealso{
\code{\link{plot2base64}}
}
