% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_report.R
\name{build_limma_report}
\alias{build_limma_report}
\title{Build Cluster Hits Report}
\usage{
build_limma_report(
  header_section,
  plots,
  plots_sizes,
  level_headers_info,
  output_file_path
)
}
\arguments{
\item{header_section}{A character string containing the HTML header section.}

\item{plots}{A list of ggplot2 plot objects.}

\item{plots_sizes}{A list of integers specifying the size of each plot.}

\item{level_headers_info}{A list of header information for each level.}

\item{output_file_path}{A character string specifying the path to save the 
HTML report.}

\item{spline_params}{A list of spline parameters.}

\item{mode}{A character string specifying the mode 
('isolated' or 'integrated').}
}
\value{
No return value, called for side effects.
}
\description{
Generates an HTML report for clustered hits, including plots and 
spline parameter details, with a table of contents.
}
\seealso{
\code{\link{plot2base64}}, \code{\link{create_progress_bar}}
}
