% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{calc_cv}
\alias{calc_cv}
\title{Calculate average CV across unique time points}
\usage{
calc_cv(time_values, response_values)
}
\arguments{
\item{time_values}{A numeric vector containing the time points. Time points
may repeat across replicates.}

\item{response_values}{A numeric vector of response values corresponding to
the time points.}
}
\value{
The average coefficient of variation (CV) across all time points.
Returns NA if all CVs are NA.
}
\description{
This function calculates the coefficient of variation (CV) for each unique
time point based on the provided time values and response values. It then
returns the average CV across all time points. The CV is only calculated if
there are more than one valid (non-NA) values for a given time point and
the mean of the values is non-zero.
}
