% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{calculate_abs_interval_diff}
\alias{calculate_abs_interval_diff}
\title{Calculate Mean of Absolute Differences and Update DataFrame by Level}
\usage{
calculate_abs_interval_diff(all_levels_clustering, data, meta, condition)
}
\arguments{
\item{all_levels_clustering}{A list of clustering levels, each containing
a dataframe with a `top_table` element.}

\item{data}{A dataframe containing only numeric values.}

\item{meta}{A dataframe containing meta information about `data` with a 
column `Time` for biological replicates.}

\item{condition}{A string specifying the column name in `meta` that divides 
the experiment into levels.}
}
\value{
The updated `all_levels_clustering` with a new column 
`abs_interval_diff` in each `top_table` for the corresponding level.
}
\description{
This function takes `all_levels_clustering`, `data`, `meta`, and `condition` 
as inputs, calculates the mean of the absolute difference between consecutive
averages of biological replicates for each feature (row) in `data` over time, 
and updates `all_levels_clustering` with these mean values based on 
`feature_index`.
}
