% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gsea_report.R
\name{check_clustered_hits}
\alias{check_clustered_hits}
\title{Check Clustered Genes Dataframe for Required Conditions}
\usage{
check_clustered_hits(levels_clustered_hits)
}
\arguments{
\item{levels_clustered_hits}{A list of dataframes to be checked for the
required format.}
}
\value{
This function does not return a value. It stops with an error message
        if the conditions are not met.
}
\description{
This function checks if a given dataframe `clustered_genes` contains the
required columns `gene` and `cluster`. The `gene` column must contain only
character strings of length 1, and the `cluster` column must contain only
integers. If any condition is not met, the function stops the script and
produces an informative error message.
}
