% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{check_colums_spline_test_configs}
\alias{check_colums_spline_test_configs}
\title{Check Columns in Spline Test Configurations}
\usage{
check_colums_spline_test_configs(spline_test_configs)
}
\arguments{
\item{spline_test_configs}{A dataframe containing spline test configurations.}
}
\value{
No return value, called for side effects.
}
\description{
Validates that the spline test configurations contain the required columns 
in the correct order.
}
\examples{
spline_test_configs <- data.frame(
  spline_type = c("type1", "type2"), 
  degree = c(3, 4), 
  dof = c(5, 6), 
  knots = c(1, 2), 
  bknots = c(0.5, 1.5)
)
check_colums_spline_test_configs(spline_test_configs)

}
