% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_data}
\alias{check_data}
\title{Check Data Matrix}
\usage{
check_data(data, data_meta_index = NA)
}
\arguments{
\item{data}{A matrix containing numeric values.}

\item{data_meta_index}{An optional parameter specifying the index of the data 
for error messages. Default is NA.}
}
\value{
Returns TRUE if all checks pass. Stops execution and returns an error
 message if any check fails.
}
\description{
This function checks the validity of the data matrix, ensuring that it is a 
matrix, contains only numeric values, 
has no missing values, and all elements are non-negative. Additionally, it
 verifies that no rows or columns are 
entirely zeros.
}
\examples{
# Example of how to use the function
data <- matrix(runif(9), nrow = 3, ncol = 3)
check_data(data)

invalid_data <- matrix(c(1, 2, NA, 4, 5, 6, 7, 8, 9), nrow = 3, ncol = 3)
try(check_data(invalid_data))  # Should raise an error

}
