% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_data_and_meta}
\alias{check_data_and_meta}
\title{utils scripts contains shared functions that are used by at least two package 
functions of the SplineOmics package.
Check Data and Meta}
\usage{
check_data_and_meta(
  data,
  meta,
  condition,
  meta_batch_column = NA,
  data_meta_index = NA
)
}
\arguments{
\item{data}{A matrix containing numeric values.}

\item{meta}{A dataframe containing the metadata, including the 'Time' column 
and the specified condition column.}

\item{condition}{A single character string specifying the column name in 
the meta dataframe to be checked.}

\item{meta_batch_column}{An optional parameter specifying the column name 
in the meta dataframe used to remove the batch effect. Default is NA.}

\item{data_meta_index}{An optional parameter specifying the index of the 
data/meta pair for error messages. Default is NA.}
}
\value{
Returns TRUE if all checks pass. Stops execution and returns an 
error message if any check fails.
}
\description{
This function checks the validity of the data and meta objects, ensuring that 
data is a matrix with numeric values and that meta is a dataframe containing 
the specified condition column. Additionally, it verifies that the number of 
columns in the data matrix matches the number of rows in the meta dataframe.
}
\examples{
# Example of how to use the function
data <- matrix(runif(9), nrow = 3, ncol = 3)
meta <- data.frame(Time = 1:3, Condition = c("A", "B", "C"))
check_data_and_meta(data, meta, "Condition")

}
