% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{check_databases}
\alias{check_databases}
\title{Check Valid Databases Dataframe}
\usage{
check_databases(databases)
}
\arguments{
\item{databases}{A dataframe to check.}
}
\value{
None. This function stops execution and provides an
        error message if the dataframe is not valid.
}
\description{
This function checks if the dataframe has exactly three columns
named DB, Geneset, and Gene, and all columns must be of type
character.
}
