% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{check_designs}
\alias{check_designs}
\title{Check Designs}
\usage{
check_designs(designs)
}
\arguments{
\item{designs}{A character vector of design descriptions.}
}
\value{
No return value, called for side effects.
}
\description{
Validates that the design descriptions are character vectors with each 
element not exceeding 75 characters in length and containing 'X'.
}
\examples{
check_designs(designs = c("X_variable1", "X_variable2"))

}
\seealso{
\code{\link{stop}} for error handling.
}
