% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{check_meta}
\alias{check_meta}
\title{Check Meta}
\usage{
check_meta(meta, meta_batch_column)
}
\arguments{
\item{meta}{A dataframe containing metadata.}

\item{meta_batch_column}{A character string specifying the meta batch column.}
}
\value{
No return value, called for side effects.
}
\description{
Validates the metadata dataframe ensuring it contains the required 'Time' 
column and checks the presence of the specified batch column if provided.
}
\examples{
meta <- data.frame(Time = seq(1, 10), batch = rep(c("A", "B"), each = 5))
check_meta(meta, "batch")

}
\seealso{
\code{\link{removeBatchEffect}}
}
