% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_mode}
\alias{check_mode}
\title{utils.R contains shared functions that are used by at least two package 
functions of the splinetime package.}
\usage{
check_mode(mode)
}
\description{
utils.R contains shared functions that are used by at least two package 
functions of the splinetime package.
}
