% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_null_elements}
\alias{check_null_elements}
\title{Check for NULL Elements in Arguments}
\usage{
check_null_elements(args)
}
\arguments{
\item{args}{A list of arguments to check for `NULL` elements.}
}
\value{
This function does not return a value. It stops execution if 
any `NULL` elements are found in the input list.
}
\description{
This function checks if any elements in the provided list of arguments 
are `NULL`.
If any `NULL` elements are found, it stops the execution and returns 
an informative error message.
}
