% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_padjust_method}
\alias{check_padjust_method}
\title{Check p-Adjustment Method}
\usage{
check_padjust_method(padjust_method)
}
\arguments{
\item{padjust_method}{A character string specifying the p-adjustment method. 
Valid options 
are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", and 
"none".}
}
\value{
Returns TRUE if the p-adjustment method is valid. Stops execution and 
returns an error message if the method is invalid.
}
\description{
This function checks if the provided p-adjustment method is valid. The valid 
methods are:
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", and "none". 
If the method
is not one of these, the function stops execution and returns an error
 message.
}
\examples{
# Example of how to use the function
check_padjust_method("BH")  # Should pass without error

try(check_padjust_method("invalid_method"))  # Should raise an error

}
