% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{check_params}
\alias{check_params}
\title{Check Params List for Required Conditions}
\usage{
check_params(params)
}
\arguments{
\item{params}{A list to be checked for the required conditions, or `NA`.}
}
\value{
This function does not return a value. It stops with an error message
        if the conditions are not met.
}
\description{
This function checks if a given list `params` contains only the allowed
named elements. The elements do not have to be present, but if they are,
they must be
named exactly as specified and must contain the correct data types: float,
character, int,
int, and float. If any condition is not met, the function stops the script
and produces an
informative error message. `params` can also be `NA`.
}
