% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_spline_params_generally}
\alias{check_spline_params_generally}
\title{Check Spline Parameters Generally}
\usage{
check_spline_params_generally(spline_params)
}
\arguments{
\item{spline_params}{A list of spline parameters.}
}
\value{
No return value, called for side effects.
}
\description{
Validates the general structure and contents of spline parameters.
}
\examples{
\dontrun{
spline_params <- list(
  spline_type = c("b", "n"),
  degree = c(3, NA),
  dof = c(4, NA),
  knots = list(NA, c(1, 2, 3)),
  bknots = list(NA, c(0.5, 1.5))
)
check_spline_params_generally(spline_params)}

}
\seealso{
\code{\link{check_spline_params_mode_dependent}}
}
