% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_spline_params_mode_dependent}
\alias{check_spline_params_mode_dependent}
\title{Check Spline Parameters Mode Dependent}
\usage{
check_spline_params_mode_dependent(spline_params, mode, meta, condition)
}
\arguments{
\item{spline_params}{A list of spline parameters.}

\item{mode}{A character string specifying the mode 
('integrated' or 'isolated').}

\item{meta}{A dataframe containing metadata.}

\item{condition}{A character string specifying the condition.}
}
\value{
No return value, called for side effects.
}
\description{
Validates the spline parameters depending on the specified mode.
}
\examples{
\dontrun{
spline_params <- list(
  spline_type = "b",
  degree = 3,
  dof = 4,
  knots = list(c(1, 2, 3)),
  bknots = list(c(0.5, 1.5))
)
check_spline_params_mode_dependent(spline_params, "integrated")}

}
\seealso{
\code{\link{check_spline_params_generally}}
}
