% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{check_spline_test_configs}
\alias{check_spline_test_configs}
\title{Check Spline Test Configurations}
\usage{
check_spline_test_configs(spline_test_configs, metas)
}
\arguments{
\item{spline_test_configs}{A configuration object for spline tests.}

\item{metas}{A list of metadata corresponding to the data matrices.}
}
\value{
No return value, called for side effects.
}
\description{
Validates the spline test configurations by checking required columns, 
spline type column, spline type parameters, and maximum and minimum degrees 
of freedom against the metadata.
}
\seealso{
\code{\link{check_colums_spline_test_configs}}, 
\code{\link{check_spline_type_column}}, 
\code{\link{check_spline_type_params}}, 
\code{\link{check_max_and_min_dof}}
}
