% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{check_spline_type_column}
\alias{check_spline_type_column}
\title{Check Spline Type Column}
\usage{
check_spline_type_column(spline_test_configs)
}
\arguments{
\item{spline_test_configs}{A dataframe containing spline test configurations.}
}
\value{
No return value, called for side effects.
}
\description{
Validates that the 'spline_type' column in the spline test configurations 
contains only 'n' or 'b'.
}
\examples{
spline_test_configs <- data.frame(
  spline_type = c("n", "b", "n", "b"), 
  degree = c(3, 4, 5, 6), 
  dof = c(5, 6, 7, 8), 
  knots = c(1, 2, 3, 4), 
  bknots = c(0.5, 1.5, 2.5, 3.5)
)
check_spline_type_column(spline_test_configs)

}
