% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_splineomics_elements}
\alias{check_splineomics_elements}
\title{Check for Required Elements in the SplineOmics Object}
\usage{
check_splineomics_elements(splineomics, func_type)
}
\arguments{
\item{splineomics}{The object to be checked.}

\item{func_type}{A string specifying the function type. It can be one of
"cluster_hits", "create_limma_report", "run_limma_splines", or "explore_data"}
}
\value{
None. Stops execution if any required element is missing.
}
\description{
This function checks if the given object contains all the required named
elements for a specified function type. If any element is missing, it stops
the script and provides an informative error message.
}
