% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_time_unit}
\alias{check_time_unit}
\title{Check Time Unit}
\usage{
check_time_unit(time_unit)
}
\arguments{
\item{time_unit}{A character string specifying the time unit. Valid options 
are 
's' for seconds, 'm' for minutes, 'h' for hours, and 'd' for days.}
}
\value{
Returns TRUE if the time unit is valid. Stops execution and returns 
an error message if the time unit is invalid.
}
\description{
This function checks if the provided time unit is valid. The valid time 
units are:
's' for seconds, 'm' for minutes, 'h' for hours, and 'd' for days. If the 
time unit 
is not one of these, the function stops execution and returns an error 
message.
}
\examples{
# Example of how to use the function
check_time_unit("m")  # Should pass without error

try(check_time_unit("x"))  # Should raise an error

}
