% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{check_top_tables}
\alias{check_top_tables}
\title{Check Top Tables}
\usage{
check_top_tables(top_tables)
}
\arguments{
\item{top_tables}{A list of top tables from limma analysis.}
}
\value{
No return value, called for side effects.
}
\description{
Validates that the top tables are a list of dataframes and checks each 
dataframe using the `check_dataframe` function.
}
\examples{
\dontrun{
top_tables <- list(data.frame(feature_index = 1:10, adj.P.Val = runif(10)))
check_top_tables(top_tables)}

}
\seealso{
\code{\link{check_dataframe}}
}
