% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{control_inputs_create_gsea_report}
\alias{control_inputs_create_gsea_report}
\title{Control Inputs for GSEA Report}
\usage{
control_inputs_create_gsea_report(
  levels_clustered_hits,
  databases,
  params,
  plot_titles,
  background
)
}
\arguments{
\item{levels_clustered_hits}{A list containing clustered hits at various 
levels.}

\item{databases}{A list of databases to be used in the GSEA analysis.}

\item{params}{A list of parameters for the GSEA analysis.}

\item{plot_titles}{A character vector of titles for the plots, with length 
matching `levels_clustered_hits`.}

\item{background}{A character vector of background genes or NULL.}
}
\description{
Validates the inputs for generating a GSEA report, including clustered 
hits, genes, databases, parameters, plot titles, and background genes.
}
