% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{create_error_message}
\alias{create_error_message}
\title{Create Error Message}
\usage{
create_error_message(message, data_meta_index = NA)
}
\arguments{
\item{message}{A character string specifying the error message.}

\item{data_meta_index}{An optional parameter specifying the index of the 
data/meta pair for the error message. Default is NA.}
}
\value{
Returns a formatted error message string. If an index is provided, 
the message includes the index; otherwise, it returns the message as is.
}
\description{
This function creates a formatted error message that includes the index of 
the data/meta pair if provided. 
If no index is provided, it returns the message as is.
}
\examples{
# Example of how to use the function
message <- "data must be a matrix."
create_error_message(message, 1)
create_error_message(message)

}
