% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_limma_report.R
\name{create_p_value_histogram}
\alias{create_p_value_histogram}
\title{Create a p-value histogram from a limma top_table}
\usage{
create_p_value_histogram(top_table, title = "P-Value Histogram")
}
\arguments{
\item{top_table}{A data frame containing the limma top_table with
a column named `P.Value` for unadjusted p-values.}

\item{title}{A character string for the title of the histogram.}
}
\value{
A ggplot2 object representing the histogram of unadjusted p-values.
}
\description{
This function generates a histogram of the unadjusted p-values from a
limma top_table.
}
