% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{create_progress_bar}
\alias{create_progress_bar}
\title{Create Progress Bar}
\usage{
create_progress_bar(iterable)
}
\arguments{
\item{iterable}{An iterable object (e.g., list or vector) whose length 
determines the total number of steps.}
}
\value{
A progress bar object from the 'progress' package.
}
\description{
Creates a progress bar for tracking the progress of an iterable task.
}
\examples{
\dontrun{
items <- 1:10
pb <- create_progress_bar(items)
for (i in items) {
  pb$tick()
  Sys.sleep(0.1)
}}

}
\seealso{
\code{\link{progress_bar}}
}
