% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{create_toc}
\alias{create_toc}
\title{Create Table of Contents}
\usage{
create_toc()
}
\value{
A string containing the HTML for the Table of Contents.
}
\description{
Creates the HTML content for the Table of Contents.
}
