% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{define_html_styles}
\alias{define_html_styles}
\title{Define HTML Styles}
\usage{
define_html_styles()
}
\value{
A list containing the styles for section headers and TOC entries.
}
\description{
Defines the CSS styles for section headers and Table of Contents (TOC)
entries used in the GSEA report generation.
}
