% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo}
\alias{demo}
\title{Interactive Demo for Getting Started}
\usage{
demo()
}
\description{
This function opens the `demo.Rmd` file in RStudio for
interactive use. Users can then run each code chunk step by step.
}
