% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{design2design_matrix}
\alias{design2design_matrix}
\title{Create Design Matrix for Splines}
\usage{
design2design_matrix(meta, spline_params, level_index, design)
}
\arguments{
\item{meta}{A dataframe containing the metadata, including the time column.}

\item{spline_params}{A list containing the spline parameters. This list can
include `dof` (degrees of freedom), `knots`, `bknots` (boundary knots),
`spline_type`, and `degree`.}

\item{level_index}{An integer representing the current level index for which
the design matrix is being generated.}

\item{design}{A character string representing the design formula to be used
for generating the model matrix.}
}
\value{
A design matrix constructed using the specified spline parameters and
design formula.
}
\description{
This function generates a design matrix using spline parameters and metadata.
It accommodates both B-splines and natural cubic splines based on the provided
spline type and parameters.
}
