% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{determine_analysis_mode}
\alias{determine_analysis_mode}
\title{Determine Analysis Mode Based on Design Formula}
\usage{
determine_analysis_mode(design, factor_column)
}
\arguments{
\item{design}{A character string representing the design formula to be used 
for generating the model matrix.}

\item{factor_column}{A character string representing the column name of the 
factor of the experiment in the metadata.}
}
\value{
A character string indicating the analysis mode, either "integrated" 
if the design formula involves interaction terms with the factor of the 
experiment, or "isolated" otherwise.
}
\description{
This function determines whether each level should be analyzed in isolation 
or together based on the design formula. If the design formula includes 
interaction terms involving the factor of the experiment, the analysis 
mode is considered integrated (together). Otherwise, it is considered isolated.
}
\keyword{internal}
