% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{encode_df_to_base64}
\alias{encode_df_to_base64}
\title{Encode DataFrame to Base64 for HTML Embedding}
\usage{
encode_df_to_base64(df)
}
\arguments{
\item{df}{A dataframe to be encoded.}
}
\value{
A character string containing the base64 encoded CSV data.
}
\description{
This function takes a dataframe as input and returns a base64 encoded 
CSV object. The encoded object can be embedded into an HTML document 
directly, with a button to download the file without pointing to a 
local file.
}
