% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{ensure_clusterProfiler}
\alias{ensure_clusterProfiler}
\title{Ensure 'clusterProfiler' is installed and loaded}
\usage{
ensure_clusterProfiler()
}
\description{
This function checks if the 'clusterProfiler' package is installed.
If not, it prompts the user to choose whether to install it automatically,
install it manually, or cancel the operation. Once installed, the package
is loaded for use.
}
