% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{format_text}
\alias{format_text}
\title{Format text}
\usage{
format_text(text)
}
\arguments{
\item{text}{A character vector to be formatted.}
}
\value{
A character vector with formatted text containing line breaks.
}
\description{
This function takes a character vector `text` and splits it into individual
characters. It then iterates over the characters and builds lines not 
exceeding
a specified character limit (default 70). Newlines are inserted between lines
using the `<br>` tag, suitable for HTML display.
}
