% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{generate_and_write_html}
\alias{generate_and_write_html}
\title{Generate and Write HTML Content with Table of Contents}
\usage{
generate_and_write_html(toc, html_content, output_file_path)
}
\arguments{
\item{toc}{The HTML string representing the table of contents.}

\item{html_content}{The initial HTML content containing a placeholder for the TOC.}

\item{output_file_path}{The file path where the final HTML content will be written.}
}
\description{
This function generates HTML content by inserting a table of contents (TOC) 
at a specified placeholder and writes the final HTML content to an output file.
}
