% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{generate_explore_plots}
\alias{generate_explore_plots}
\title{Generate exploratory plots}
\usage{
generate_explore_plots(data, meta, condition)
}
\arguments{
\item{data}{A data frame or matrix containing the data to be plotted.}

\item{meta}{A data frame containing metadata associated with the data.}

\item{condition}{A string specifying the column in the metadata that contains
the condition or grouping variable.}
}
\value{
A list containing two elements:
\describe{
  \item{plots}{A list of ggplot objects representing the generated plots.}
  \item{plots_sizes}{A vector of numeric values indicating the sizes of the
  corresponding plots.}
}
}
\description{
This function generates various exploratory plots including density plots,
box plots, violin plots, PCA plots, and correlation heatmaps based on the
provided data and metadata.
}
