% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{generate_feature_name_template}
\alias{generate_feature_name_template}
\title{Generate Column Headers String from Feature Name}
\usage{
generate_feature_name_template(df)
}
\arguments{
\item{df}{A dataframe containing a column called `feature_name`.}
}
\value{
A string in the format of the value in `feature_name` but containing
the column headers instead of the values.
}
\description{
This function takes a dataframe as input and returns a string that replaces the values
in the `feature_name` column of the first row with their respective column headers.
}
