% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_limma_report.R
\name{generate_interaction_plots}
\alias{generate_interaction_plots}
\title{Generate Plots for Interaction of Condition and Time}
\usage{
generate_interaction_plots(interaction_condition_time, adj_pthresh)
}
\arguments{
\item{interaction_condition_time}{A list of top tables from the LIMMA 
analysis 
representing the interaction effects between condition and time.}

\item{adj_pthresh}{A numeric value specifying the adjusted p-value threshold 
for significance.}
}
\value{
A list containing the plots and their sizes, as well as the 
section header information.
}
\description{
Creates p-value histograms for each interaction condition in the 
interaction of condition and time. This function is used internally in the 
`create_limma_report` function.
}
