% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{generate_report_html}
\alias{generate_report_html}
\title{Generate HTML Report with Plot List}
\usage{
generate_report_html(plot_list, plot_list_nrows, report_dir)
}
\arguments{
\item{plot_list}{A list of ggplot objects to be included in the report.}

\item{plot_list_nrows}{The number of rows to arrange the plots in the report.}

\item{report_dir}{The directory where the report HTML file will be saved.}
}
\description{
This function generates an HTML report for a given list of plots. It 
includes
headers and design details specific to the data type being analyzed, and 
saves
the report in the specified directory. The function is internal and not 
exported.
}
