% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinetime_general_fun.R
\name{generate_report_html_new}
\alias{generate_report_html_new}
\title{Make dynamic HTML text descriptions, such as header- and design_text.
Future task}
\usage{
generate_report_html_new(plots, plot_len, report_dir, filename, timestamp)
}
\description{
Make dynamic HTML text descriptions, such as header- and design_text.
Future task
}
