% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gsea_report.R
\name{generate_section_content}
\alias{generate_section_content}
\title{Generate Section Content}
\usage{
generate_section_content(
  section_info,
  index,
  toc,
  html_content,
  section_header_style,
  toc_style
)
}
\arguments{
\item{section_info}{A list containing the information for the section.}

\item{index}{The index of the current section.}

\item{toc}{The current state of the Table of Contents.}

\item{html_content}{The current state of the HTML content.}

\item{section_header_style}{The CSS style for the section headers.}

\item{toc_style}{The CSS style for the TOC entries.}
}
\value{
A list with updated HTML content and TOC.
}
\description{
Generates the HTML content for a section, including headers and enrichment
results.
}
