% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{generate_spline_comparisons}
\alias{generate_spline_comparisons}
\title{Generate spline comparison plots for all condition pairs}
\usage{
generate_spline_comparisons(
  splineomics,
  all_levels_clustering,
  data,
  meta,
  condition,
  plot_info,
  adj_pthresh_avrg_diff_conditions,
  adj_pthresh_interaction
)
}
\arguments{
\item{splineomics}{A list containing the splineomics results, including
 time effects, 
average difference between conditions, and interaction between condition 
and time.}

\item{all_levels_clustering}{A list containing the X matrices for each 
condition, used 
for spline fitting.}

\item{data}{The data matrix containing the measurements.}

\item{meta}{The metadata associated with the measurements, which includes
the condition.}

\item{condition}{Column name of meta that contains the levels of the
experiment.}

\item{plot_info}{A list containing plotting information such as time unit 
and axis labels.}

\item{adj_pthresh_avrg_diff_conditions}{The adjusted p-value threshold for
 the average 
difference between conditions.}

\item{adj_pthresh_interaction}{The adjusted p-value threshold for the 
interaction 
between condition and time.}
}
\value{
A list of lists containing the comparison plots and feature names
        for each condition pair.
}
\description{
This function generates spline comparison plots for all pairwise 
combinations of conditions in the metadata. For each condition pair, it 
compares the time effects of two conditions, plots the data points, and 
overlays the fitted spline curves. The function only generates plots if 
the adjusted p-values for the average difference between conditions and the 
interaction between condition and time are below the specified thresholds.
}
