% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{get_category_2_and_3_hits}
\alias{get_category_2_and_3_hits}
\title{Get Category 2 and 3 Hits}
\usage{
get_category_2_and_3_hits(
  splineomics,
  adj_pthresh_avrg_diff_conditions,
  adj_pthresh_interaction
)
}
\arguments{
\item{splineomics}{An S3 object containing the limma top tables. It must
include the elements `avrg_diff_conditions` and `interaction_condition_time`,
which are lists of dataframes.}

\item{adj_pthresh_avrg_diff_conditions}{Numeric. Threshold for adjusted
p-value for average differences between conditions.}

\item{adj_pthresh_interaction}{Numeric. Threshold for adjusted p-value for
the interaction between condition and time.}
}
\value{
A list with two elements:
\itemize{
  \item \code{category_2_hits}: A list of filtered dataframes from
  `avrg_diff_conditions`.
  \item \code{category_3_hits}: A list of filtered dataframes from
  `interaction_condition_time`.
}
}
\description{
This function filters the `limma` top tables in the `splineomics` object
based on adjusted p-value thresholds for the average difference between
conditions (`adj_pthresh_avrg_diff_conditions`) and the interaction between
condition and time (`adj_pthresh_interaction`).
}
\examples{
category_2_and_3_hits <- get_category_2_and_3_hits(
  splineomics = splineomics,
  adj_pthresh_avrg_diff_conditions = 0.05,
  adj_pthresh_interaction = 0.01
)
}
