% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{get_curve_values}
\alias{get_curve_values}
\title{Calculate Curve Values Based on Top Table Filter}
\usage{
get_curve_values(top_table, p_value, level, meta, condition)
}
\arguments{
\item{top_table}{A data frame containing data with a column for adjusted 
p-values and 
                 expression averages which indicate the number of degrees 
                 of freedom.}

\item{p_value}{The threshold for filtering entries based on adjusted
p-values.}

\item{level}{The specific level of the condition to filter on in the 
metadata.}

\item{meta}{Metadata containing time points and conditions.}

\item{condition}{The name of the condition column in the metadata to filter 
on.}
}
\value{
A list containing two elements: `curve_values`, a data frame of 
curve values for each filtered entry, and `smooth_timepoints`, the time
points at which curves were evaluated.
}
\description{
This function filters entries from a given top table based on an adjusted 
p-value threshold,
performs spline interpolation using specified degrees of freedom, and 
calculates curve values 
for the selected entries at predefined time points. The function is 
internal and not exported.
}
