% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{get_header_section}
\alias{get_header_section}
\title{Get Header Section}
\usage{
get_header_section(title, header_text, report_type)
}
\arguments{
\item{title}{A string specifying the title of the HTML document.}

\item{header_text}{A string specifying the text to be displayed in the 
header of the report.}

\item{report_type}{A character specifying the type of HTML report.}
}
\value{
A string containing the HTML header section.
}
\description{
Generates the HTML header section for a report, including the title, header 
text, and logo. This section also includes the styling for the table and 
other HTML elements.
}
\details{
The function checks the `DEVTOOLS_LOAD` environment variable to determine 
the path to the logo image. The logo image is then converted to a base64 
data URI and included in the HTML. The header section includes styles for 
tables, table cells, and header elements to ensure proper formatting and 
alignment.
}
