% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{get_hyperparams_screen_plots_explanations}
\alias{get_hyperparams_screen_plots_explanations}
\title{Get Explanations for Hyperparameter Screen Plots}
\usage{
get_hyperparams_screen_plots_explanations()
}
\description{
This function provides textual explanations for various hyperparameter
screen plots used in the analysis. These explanations include descriptions
of the 'Venn-Heatmap' and bar plots, as well as details on the spline
curves shown for hits and non-significant features.
}
