% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{get_spline_params_info}
\alias{get_spline_params_info}
\title{Get Spline Parameters Info}
\usage{
get_spline_params_info(spline_params, j)
}
\arguments{
\item{spline_params}{A list containing the spline parameters. The list should
include elements: `spline_type`, `degree`, `dof`,
`knots`, and `bknots`.}

\item{j}{An integer specifying the index of the spline parameters to
retrieve.}
}
\value{
A character string containing HTML-formatted information about the
        spline parameters at the specified index.
}
\description{
This function retrieves the spline parameters information for a given index.
It ensures the spline parameters are valid and constructs an HTML string
describing the spline parameters.
}
\details{
The function checks if the spline parameters are not `NULL` and have a length
greater than or equal to the specified index `j`. If a parameter is
invalid or
missing, it sets the parameter to `NA`. It then constructs an HTML string
describing the spline parameters, including spline type, degree, degrees of
freedom (DoF), knots, and boundary knots.
}
