% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{hc_add}
\alias{hc_add}
\title{Add Data to Hit Comparison Object}
\usage{
hc_add(hc_obj, top_table, params_id, condition = 1, threshold = 0.05)
}
\arguments{
\item{hc_obj}{An object of class "hitcomp".}

\item{top_table}{A dataframe containing the top table data.}

\item{params_id}{A character string identifying the parameters
(max length 70).}

\item{condition}{An integer (1 or 2) specifying the condition to which the
data belongs.}

\item{threshold}{A numeric value specifying the adjusted p-value threshold.}
}
\value{
The updated hit comparison object.
}
\description{
Adds a new entry to the hit comparison object for a specified condition.
}
