% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{hc_new}
\alias{hc_new}
\title{Create New Hit Comparison Object}
\usage{
hc_new(cond1name = "Condition 1", cond2name = "Condition 2")
}
\arguments{
\item{cond1name}{A character string for the first condition name 
(max length 25).}

\item{cond2name}{A character string for the second condition name 
(max length 25).}
}
\value{
An object of class "hitcomp" containing empty data lists 
and condition names.
}
\description{
Creates a new hit comparison object with specified condition names.
}
