% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{hc_vennheatmap}
\alias{hc_vennheatmap}
\title{Generate Venn Heatmap}
\usage{
hc_vennheatmap(hc_obj)
}
\arguments{
\item{hc_obj}{An object of class "hitcomp" containing hit data for two 
conditions.}
}
\value{
A list containing the Venn heatmap plot and the number of hits.
}
\description{
Creates a Venn heatmap to visualize the overlap of hits between two 
conditions stored in a hit comparison object.
}
\seealso{
\code{\link{store_hits}}, \link[pheatmap]{pheatmap}
}
