% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{hierarchical_clustering}
\alias{hierarchical_clustering}
\title{Hierarchical Clustering of Curve Values}
\usage{
hierarchical_clustering(curve_values, k, smooth_timepoints, top_table)
}
\arguments{
\item{curve_values}{A matrix or data frame of curve values to cluster.}

\item{k}{The number of clusters to use.}

\item{smooth_timepoints}{Numeric vector of time points corresponding to
columns in curve_values.}

\item{top_table}{Data frame to be updated with cluster assignments.}
}
\value{
A list containing clustering results and the modified top_table.
}
\description{
Performs hierarchical clustering on given curve values.
The function adjusts the provided top_table with cluster
assignments.
}
