% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{huge_table_user_prompter}
\alias{huge_table_user_prompter}
\title{Check if any table in a list has more than 300 rows and prompt user for
input.}
\usage{
huge_table_user_prompter(tables)
}
\arguments{
\item{tables}{A list of data frames.}
}
\value{
NULL. This function is used for its side effects (prompting the
user and potentially stopping the script).
}
\description{
This function iterates over a list of tables and checks if any table has
 more than 300 rows.
If such a table is found, it prompts the user to proceed or stop.
}
