% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{interactive_demo}
\alias{interactive_demo}
\title{Interactive Demo for Getting Started}
\usage{
interactive_demo()
}
\description{
This function opens the `get-started.Rmd` file in RStudio for
interactive use. Users can then run each code chunk step by step.
}
