% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{is_not_na}
\alias{is_not_na}
\title{Check if Not All Values are NA}
\usage{
is_not_na(x)
}
\arguments{
\item{x}{An atomic vector or any other object.}
}
\value{
TRUE if the vector contains at least one non-NA value or if the 
object is not atomic; FALSE otherwise.
}
\description{
Determines if a given atomic vector contains at least one non-NA value.
}
