% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{Level4Functions}
\alias{Level4Functions}
\title{Level4Functions: A class for level 3 utility functions}
\description{
Level4Functions: A class for level 3 utility functions

Level4Functions: A class for level 3 utility functions
}
\details{
This class provides methods for creating error messages and checking
batch columns.

Create Error Message
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Level4Functions-create_error_message}{\code{Level4Functions$create_error_message()}}
\item \href{#method-Level4Functions-clone}{\code{Level4Functions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Level4Functions-create_error_message"></a>}}
\if{latex}{\out{\hypertarget{method-Level4Functions-create_error_message}{}}}
\subsection{Method \code{create_error_message()}}{
This method creates a formatted error message that includes the index of
the data/meta pair if provided.
If no index is provided, it returns the message as is.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Level4Functions$create_error_message(message, data_meta_index = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{A character string specifying the error message.}

\item{\code{data_meta_index}}{An optional parameter specifying the index of the
data/meta pair for the error message. Default is NA.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a formatted error message string. If an index is provided,
the message includes the index; otherwise, it returns the message as is.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Level4Functions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Level4Functions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Level4Functions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
