% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_box_plots}
\alias{make_box_plots}
\title{Generate Boxplot}
\usage{
make_box_plots(data, meta, condition)
}
\arguments{
\item{data}{A numeric matrix containing the data.}
}
\value{
A ggplot object representing the boxplot.
}
\description{
This function generates a boxplot for a given data matrix. The boxplot shows 
the distribution of the values in the data matrix across different variables, 
with log2 intensity on the y-axis.
}
\examples{
\dontrun{
# Example usage
data <- matrix(rnorm(1000), ncol = 10)
boxplot <- make_boxplot(data)
print(boxplot)}

}
