% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_correlation_heatmaps}
\alias{make_correlation_heatmaps}
\title{Generate Correlation Heatmaps}
\usage{
make_correlation_heatmaps(data, meta, condition)
}
\arguments{
\item{data}{A numeric matrix containing the data.}

\item{meta}{A dataframe containing the metadata.}

\item{condition}{The column name in the metadata dataframe that contains the 
factor levels for generating individual heatmaps.}
}
\value{
A list of `ComplexHeatmap` heatmap objects representing the 
correlation 
heatmaps.
}
\description{
This function generates correlation heatmaps using Spearman correlation for 
a given data matrix. It creates a combined heatmap for all levels and 
individual heatmaps for each level specified in the condition column of the 
metadata.
}
