% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{make_enrich_dotplot}
\alias{make_enrich_dotplot}
\title{Make Enrich Dotplot}
\usage{
make_enrich_dotplot(enrichments_list, databases, title = "Title")
}
\arguments{
\item{enrichments_list}{A list of enrichments containing data frames for 
different databases.}

\item{databases}{A character vector specifying the databases to be included.}

\item{title}{A character string specifying the title of the dotplot.}
}
\value{
A list containing:
\describe{
  \item{p}{The ggplot object representing the dotplot.}
  \item{dotplot_nrows}{An integer specifying the number of rows in the 
                       dotplot.}
  \item{full_enrich_results}{A data frame containing the full enrichments 
                             results.}
}
}
\description{
Make an enriched dotplot for visualization.
}
