% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_mds_plot}
\alias{make_mds_plot}
\title{Generate MDS Plot}
\usage{
make_mds_plot(data, meta, condition)
}
\arguments{
\item{data}{A numeric matrix containing the data.}

\item{meta}{A dataframe, containign the meta information of data.}

\item{condition}{The column of the meta dataframe containign the levels that 
separate the experiment.}
}
\value{
A ggplot object representing the MDS plot.
}
\description{
This function generates a multidimensional scaling (MDS) plot for a given 
data matrix. The MDS plot visualizes the similarities or dissimilarities 
between samples in the data matrix.
}
