% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_pca_plot}
\alias{make_pca_plot}
\title{Generate PCA Plot with Dynamic Coloring}
\usage{
make_pca_plot(data, meta, condition)
}
\arguments{
\item{data}{A numeric matrix containing the data.}

\item{meta}{A dataframe containing the metadata.}

\item{condition}{The column name in the metadata dataframe that contains the
factor levels for coloring the PCA plot.}
}
\value{
A ggplot object representing the PCA plot.
}
\description{
This function generates a PCA plot from a data matrix, dynamically coloring
the points based on the levels of a specified factor in the metadata.
}
