% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_pca_variance_explained_plot}
\alias{make_pca_variance_explained_plot}
\title{Generate PCA Variance Explained Plot}
\usage{
make_pca_variance_explained_plot(data, meta, condition)
}
\arguments{
\item{data}{A numeric matrix containing the data.}

\item{meta}{A dataframe, containign the meta information of data.}

\item{condition}{The column of the meta dataframe containign the levels that 
separate the experiment.}
}
\value{
A ggplot object representing the variance explained by the principal 
components.
}
\description{
This function generates a bar plot showing the variance explained by the 
first seven principal components of a given data matrix.
}
