% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_tsne_plot}
\alias{make_tsne_plot}
\title{Generate a t-SNE Plot}
\usage{
make_tsne_plot(data, meta, condition)
}
\arguments{
\item{data}{A matrix where rows represent features and columns represent 
samples.}

\item{meta}{A data frame containing metadata for the samples, with rows 
corresponding to columns in the data matrix.}

\item{condition}{The name of the column in the metadata data frame that 
contains the condition labels for coloring the plot.}
}
\value{
A ggplot2 object representing the t-SNE plot.
}
\description{
This function performs t-SNE on a given data matrix and creates a ggplot2 
scatter plot where points are colored based on a specified condition from 
the metadata.
}
