% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_violin_plots}
\alias{make_violin_plots}
\title{Generate Violin Plot}
\usage{
make_violin_plots(data, meta, condition)
}
\arguments{
\item{data}{A numeric matrix containing the data.}
}
\value{
A ggplot object representing the violin plot.
}
\description{
This function generates a violin plot for a given data matrix. The violin 
plot shows the distribution of the values in the data matrix across different 
variables, with each variable's distribution displayed as a separate violin.
}
\examples{
\dontrun{
# Example usage
data <- matrix(rnorm(1000), ncol = 10)
violin_plot <- make_violin_plot(data)
print(violin_plot)}

}
