% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{maybe_add_dashed_lines}
\alias{maybe_add_dashed_lines}
\title{Conditionally add dashed lines for treatment timepoints}
\usage{
maybe_add_dashed_lines(
  p,
  plot_info,
  level,
  y_pos = 1,
  horizontal_labels = FALSE
)
}
\arguments{
\item{p}{A ggplot object. The plot to which dashed lines and labels
will be added.}

\item{plot_info}{A list containing the treatment timepoints and
treatment labels. Treatment timepoints and labels can either be
unnamed elements or named lists where each element corresponds
to a different `level`.}

\item{level}{A character string. Used to extract the treatment
timepoints and labels when they are stored in named lists.}

\item{y_pos}{A numeric value specifying the y-axis position where
the text labels should be placed. Defaults to 1.}

\item{horizontal_labels}{Boolean flag indicating whether to have a vertical
label (default) or horizontal label.}
}
\value{
A list containing:
- `p`: The ggplot object with possibly added dashed lines and labels.
- `treatment_colors`: A named vector of colors used for the treatment labels.
}
\description{
This internal function checks whether there are valid treatment
timepoints and labels in the `plot_info` list. If found, it adds
dashed vertical lines and their corresponding x-axis values to the plot.
The treatment timepoints and labels can either be named lists (for
multiple levels) or unnamed single elements.
}
