% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{merge_annotation_all_levels_clustering}
\alias{merge_annotation_all_levels_clustering}
\title{Merge Annotation with All Top Tables}
\usage{
merge_annotation_all_levels_clustering(
  all_levels_clustering,
  annotation = NULL
)
}
\arguments{
\item{all_levels_clustering}{A list where each element contains a `top_table`
dataframe with a `feature_nr` column. Some elements may be logical values.}

\item{annotation}{A dataframe containing the annotation information.}
}
\value{
A list with updated `top_table` dataframes containing merged
annotation information.
}
\description{
This function merges annotation information into the `top_table` of each
non-logical element in a list.
}
