% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{merge_top_table_with_annotation}
\alias{merge_top_table_with_annotation}
\title{Merge Annotation with a Single Top Table}
\usage{
merge_top_table_with_annotation(top_table, annotation)
}
\arguments{
\item{top_table}{A dataframe containing the `top_table` with a `feature_nr` column.}

\item{annotation}{A dataframe containing the annotation information.}
}
\value{
A dataframe with updated `top_table` containing merged annotation information.
}
\description{
This function merges annotation information into a single `top_table`
dataframe based on the `feature_nr` column.
}
