% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{modify_limma_top_table}
\alias{modify_limma_top_table}
\title{Converts top_table to a tibble and adds new column feature_name}
\usage{
modify_limma_top_table(top_table, feature_names)
}
\description{
Converts top_table to a tibble and adds new column feature_name
}
