% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{NumericBlockFinder}
\alias{NumericBlockFinder}
\title{NumericBlockFinder: A class for finding numeric blocks in data}
\description{
This class provides methods to identify the upper-left and lower-right
cells of a numeric block within a dataframe.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A dataframe containing the input data.}

\item{\code{upper_left_cell}}{A list containing the row and column indices of the
                       upper-left cell.

Initialize a NumericBlockFinder object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NumericBlockFinder-new}{\code{NumericBlockFinder$new()}}
\item \href{#method-NumericBlockFinder-find_upper_left_cell}{\code{NumericBlockFinder$find_upper_left_cell()}}
\item \href{#method-NumericBlockFinder-find_lower_right_cell}{\code{NumericBlockFinder$find_lower_right_cell()}}
\item \href{#method-NumericBlockFinder-clone}{\code{NumericBlockFinder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericBlockFinder-new"></a>}}
\if{latex}{\out{\hypertarget{method-NumericBlockFinder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumericBlockFinder$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the input data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the NumericBlockFinder class.

Find the upper-left cell of the first 6x6 block of numeric values

This method identifies the upper-left cell of the first 6x6 block of
numeric values in the dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericBlockFinder-find_upper_left_cell"></a>}}
\if{latex}{\out{\hypertarget{method-NumericBlockFinder-find_upper_left_cell}{}}}
\subsection{Method \code{find_upper_left_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumericBlockFinder$find_upper_left_cell()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing the row and column indices of the upper-left
        cell.

Find the lower-right cell of a block of contiguous non-NA values

This method identifies the lower-right cell of a block of contiguous
non-NA values starting from a given upper-left cell in the dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericBlockFinder-find_lower_right_cell"></a>}}
\if{latex}{\out{\hypertarget{method-NumericBlockFinder-find_lower_right_cell}{}}}
\subsection{Method \code{find_lower_right_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumericBlockFinder$find_lower_right_cell()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing the row and column indices of the lower-right
        cell.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumericBlockFinder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NumericBlockFinder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumericBlockFinder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
