% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tutorial&template.R
\name{open_template}
\alias{open_template}
\title{Open Template for Quick Setup}
\usage{
open_template()
}
\description{
This function opens the `template.Rmd` file in RStudio for
interactive use. The template file provides a structure for users 
to quickly set up their personal analysis.
}
