% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{perform_clustering}
\alias{perform_clustering}
\title{Perform Clustering}
\usage{
perform_clustering(top_tables, clusters, meta, condition, spline_params, mode)
}
\arguments{
\item{top_tables}{A list of top tables from limma analysis.}

\item{clusters}{A list specifying clusters.}

\item{meta}{A dataframe containing metadata.}

\item{condition}{A character string specifying the condition.}

\item{spline_params}{A list of spline parameters for the analysis.}

\item{mode}{A character string specifying the mode
('isolated' or 'integrated').}
}
\value{
A list of clustering results for each level within the condition.
}
\description{
Performs clustering on top tables using specified p-values and clusters
for each level within a condition.
}
\seealso{
\code{\link{process_level_cluster}}
}
