% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{plot2base64}
\alias{plot2base64}
\title{Convert Plot to Base64}
\usage{
plot2base64(
  plot,
  height,
  width = 7,
  base_height_per_row = 2.5,
  units = "in",
  html_img_width = "100\%"
)
}
\arguments{
\item{plot}{A ggplot2 plot object.}

\item{height}{An integer specifying the height of the plot for correct
representation in the HTML.}

\item{width}{A numeric value specifying the width of the plot in inches.}

\item{base_height_per_row}{A numeric value specifying the base height per
row in inches.}

\item{units}{A character string specifying the units for the width and
height.}

\item{html_img_width}{A character string specifying the width of the image
in HTML.}
}
\value{
A character string containing an HTML img tag with the Base64-encoded
 plot.
}
\description{
Converts a ggplot2 plot to a Base64-encoded PNG image and returns an HTML
img tag for embedding in a report.
}
\seealso{
\link[base64enc]{dataURI}
}
