% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_all_shapes}
\alias{plot_all_shapes}
\title{Plot All Shapes}
\usage{
plot_all_shapes(curve_values, time_unit_label)
}
\arguments{
\item{curve_values}{A dataframe containing curve values and cluster
assignments.}

\item{time_unit_label}{A character string specifying the time unit label.}
}
\value{
A ggplot object representing the average curves by cluster.
}
\description{
Generates a plot of average curves for each cluster, showing min-max
normalized intensities over time.
}
\seealso{
\code{\link{ggplot2}}
}
