% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{plot_composite_splines}
\alias{plot_composite_splines}
\title{Plot Composite Splines}
\usage{
plot_composite_splines(
  data,
  meta,
  spline_test_configs,
  top_table,
  top_table_name,
  indices,
  type
)
}
\arguments{
\item{data}{A matrix of data values.}

\item{meta}{A dataframe containing metadata.}

\item{spline_test_configs}{A configuration object for spline tests.}

\item{top_table}{A dataframe containing the top table results.}

\item{top_table_name}{A character string specifying the name of the 
top table.}

\item{indices}{A vector of indices specifying which features to plot.}

\item{type}{A character string specifying the type of features ('significant' 
or 'not_significant').}
}
\value{
A list containing the composite plot and its length if plots are 
generated, FALSE otherwise.
}
\description{
Generates composite spline plots for significant and non-significant 
features based on the specified indices.
}
\examples{
data <- matrix(runif(100), nrow = 10)
meta <- data.frame(Time = seq(1, 10))
spline_test_configs <- list(spline_type = c("n"), dof = list(3), 
                            knots = list(NA), bknots = list(NA))
top_table <- data.frame(feature_index = 1:10, adj.P.Val = runif(10), 
                        AveExpr = runif(10))
indices <- 1:5
plot_composite_splines(data, meta, spline_test_configs, top_table, 
                       "TopTable1", indices, "significant")

}
\seealso{
\code{\link{splines::bs}}, \code{\link{splines::ns}}, 
\code{\link{ggplot2}}, \code{\link{patchwork}}
}
