% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_consensus_shapes}
\alias{plot_consensus_shapes}
\title{Plot Consensus Shapes}
\usage{
plot_consensus_shapes(curve_values, time_unit_label)
}
\arguments{
\item{curve_values}{A dataframe containing curve values and cluster assignments.}

\item{time_unit_label}{A character string specifying the time unit label.}
}
\value{
A list containing the composite plot of consensus shapes and its size.
}
\description{
Generates composite plots of single and consensus shapes for each cluster 
of curve values.
}
\examples{
curve_values <- data.frame(
  Time1 = runif(10), Time2 = runif(10), Time3 = runif(10), 
  cluster = rep(1:2, each = 5)
)
time_unit_label <- "[min]"
plot_consensus_shapes(curve_values, time_unit_label)

}
\seealso{
\code{\link{plot_single_and_consensus_splines}}, \code{\link{patchwork}}
}
