% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{plot_cv}
\alias{plot_cv}
\title{Coefficient of Variation (CV) Plot}
\usage{
plot_cv(data, meta, condition)
}
\arguments{
\item{data}{A data frame where rows are features and columns are samples.}

\item{meta}{A data frame with sample metadata. Must contain a column "Time"
and the condition column.}

\item{condition}{The name of the column in the meta table that contains the 
condition information.}
}
\value{
A list of ggplot2 objects, each showing the distribution of CVs for
 one condition.
}
\description{
This function takes a data frame with time series data 
(rows as features and columns as samples),
a meta table with sample information including time points and conditions, 
computes the coefficient
of variation (CV) for each feature for each condition level, and plots the 
distribution of these
CVs.
}
