% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(
  datas,
  meta,
  mode,
  condition,
  all_levels_clustering,
  time_unit_label,
  cluster_heatmap_columns
)
}
\arguments{
\item{datas}{A matrix of data values.}

\item{meta}{A dataframe containing metadata.}

\item{mode}{A character vector with length 1, specifying the type of limma
design formula (integrated for formulas with interaction effects
between the levels, isolated for formulas where each level is
analysed in isolation (no interaction effects))}

\item{condition}{A character string specifying the condition.}

\item{all_levels_clustering}{A list containing clustering results for each
level within the condition.}

\item{time_unit_label}{A character string specifying the time unit label.}

\item{cluster_heatmap_columns}{Boolean specifying wether to cluster the
columns of the heatmap or not.}
}
\value{
A list of ComplexHeatmap heatmap objects for each level.
}
\description{
Generates heatmaps for each level within a condition, showing z-scores of
log2 intensity values, split by clusters.
}
\seealso{
\link[ComplexHeatmap]{Heatmap}, \link[dplyr]{arrange}
}
