% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{plot_limma_combos_results}
\alias{plot_limma_combos_results}
\title{Plot limma Combination Results}
\usage{
plot_limma_combos_results(
  top_tables_combos,
  datas,
  metas,
  condition,
  spline_test_configs,
  meta_batch_column,
  meta_batch2_column,
  time_unit = time_unit
)
}
\arguments{
\item{top_tables_combos}{A list of top tables for each combination.}

\item{datas}{A list of matrices.}

\item{metas}{A list of metadata corresponding to the data matrices.}

\item{condition}{Meta column name that contains the levels.}

\item{spline_test_configs}{A configuration object for spline tests.}

\item{meta_batch_column}{A character string specifying the meta batch column.}

\item{meta_batch2_column}{A character string specifying the second meta batch
column.}

\item{time_unit}{A single character, such as s, m, h, or d, specifying the
time_unit that should be used for the plots (s = seconds, m = minutes,
h = hours, d = days). This single character will be converted to a string 
that is a little bit more verbose, such as sec in square brackets for s.}
}
\value{
A list of results including hit comparison plots and composite 
spline plots for each pair of combinations.
}
\description{
Generates plots for pairwise comparisons of hyperparameter combinations 
using limma results.
}
