% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_single_and_consensus_splines}
\alias{plot_single_and_consensus_splines}
\title{Plot Single and Consensus Splines}
\usage{
plot_single_and_consensus_splines(time_series_data, title, time_unit_label)
}
\arguments{
\item{time_series_data}{A dataframe or matrix with time series data.}

\item{title}{A character string specifying the title of the plot.}

\item{time_unit_label}{A character string specifying the time unit label.}
}
\value{
A ggplot object representing the single and consensus shapes.
}
\description{
Generates a plot showing individual time series shapes and their consensus 
(mean) shape.
}
\seealso{
\code{\link{ggplot2}}
}
