% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_splines}
\alias{plot_splines}
\title{Plot Splines for Features Based on Top Table Information}
\usage{
plot_splines(
  top_table,
  data,
  meta,
  X,
  time_unit_label,
  plot_info,
  adj_pthreshold,
  replicate_column
)
}
\arguments{
\item{top_table}{A dataframe containing the indices and names of features,
along with their
                 statistical metrics such as intercepts and spline
                 coefficients.}

\item{data}{A matrix or dataframe containing the raw data values for each
feature.}

\item{meta}{A dataframe containing metadata for the data, including time
points.}

\item{X}{The limma design matrix that defines the experimental conditions.}

\item{time_unit_label}{A string shown in the plots as the unit for the time,
such as min or hours.}

\item{plot_info}{List containing the elements y_axis_label (string), 
time_unit (string), treatment_labels (character vector),
treatment_timepoints (integer vector). All can also be NA. 
This list is used to add this info to the spline plots. 
time_unit is used to label the x-axis, and treatment_labels
and -timepoints are used to create vertical dashed lines,
indicating the positions of the treatments (such as 
feeding, temperature shift, etc.).}

\item{adj_pthreshold}{Double > 0 and < 1 specifying the adj. p-val threshold.}

\item{replicate_column}{String specifying the column of the meta dataframe
that contains the labels of the replicate measurents. When that is not 
given, this argument is NULL.}
}
\value{
A list containing the composite plot and the number of rows used in
the plot layout.
}
\description{
This function generates plots for each feature listed in the
top table using spline
interpolation for fitted values. It creates individual plots for each feature
and combines
them into a single composite plot. The function is internal and not exported.
}
