% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{prepare_gene_lists_for_enrichr}
\alias{prepare_gene_lists_for_enrichr}
\title{Prepare Gene Lists for Enrichr and Return as String}
\usage{
prepare_gene_lists_for_enrichr(all_levels_clustering, genes)
}
\arguments{
\item{all_levels_clustering}{A list where each element contains a dataframe
`clustered_hits` with columns `feature` and `cluster`.}

\item{genes}{A vector of gene names corresponding to the feature indices.}
}
\value{
A character vector with the formatted gene lists for each cluster.
}
\description{
This function processes the clustered hits in each element of 
`all_levels_clustering`, formats the gene names for easy copy-pasting into 
Enrichr, and returns the formatted gene lists as a string.
}
