% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare Plot Data}
\usage{
prepare_plot_data(enrichments_list, databases)
}
\arguments{
\item{enrichments_list}{A list of enrichments containing data frames for
different databases.}

\item{databases}{A character vector specifying the databases to be included.}
}
\value{
A list containing two data frames:
\describe{
  \item{top_plot_data}{A data frame containing the prepared plot data for
                         visualization of top combinations.}
  \item{full_enrich_results}{A data frame containing the full enrichments
                              results.}
}
}
\description{
This function prepares plot data for visualization based on enrichments
lists and specified databases.
}
