% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{preprocess_rna_seq_data}
\alias{preprocess_rna_seq_data}
\title{Perform default preprocessing of raw RNA-seq counts}
\usage{
preprocess_rna_seq_data(raw_counts, design_matrix, normalize_func = NULL)
}
\arguments{
\item{raw_counts}{A matrix of raw RNA-seq counts (genes as rows, samples as
columns).}

\item{design_matrix}{A design matrix used in the linear modeling, typically
specifying the experimental conditions.}

\item{normalize_func}{An optional normalization function. If provided, this 
function will be used to normalize the `DGEList` object. If not provided,
 TMM normalization (via `edgeR::calcNormFactors`) will be used by default.}
}
\value{
A `voom` object, which includes the log2-counts per million (logCPM)
 matrix and observation-specific weights.
}
\description{
This function is called when `preprocess_rna_seq` is `TRUE`. It performs the 
default preprocessing steps for raw RNA-seq counts, including creating a 
`DGEList` object, normalizing the counts, and applying the `voom` 
transformation.
}
