% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{print_info_message}
\alias{print_info_message}
\title{Print Informational Message}
\usage{
print_info_message(message_prefix, report_dir)
}
\arguments{
\item{message_prefix}{A custom message prefix to be displayed before the success message.}

\item{report_dir}{The directory where the HTML reports are located.}
}
\description{
This function prints a nicely formatted informational message with a green "Info" label.
}
