% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{process_combo_pair}
\alias{process_combo_pair}
\title{Process Combination Pair}
\usage{
process_combo_pair(
  combo_pair,
  combo_pair_name,
  report_info,
  report_dir,
  timestamp
)
}
\arguments{
\item{combo_pair}{A list containing hit comparison and composite spline
plots.}

\item{combo_pair_name}{A character string for naming the combination pair.}

\item{report_info}{An object containing report information.}

\item{report_dir}{A non-empty string specifying the report directory.}

\item{timestamp}{A timestamp to include in the report filename.}
}
\value{
No return value, called for side effects.
}
\description{
Processes a combination pair to generate plots and compile them into an
HTML report.
}
\seealso{
\code{\link{generate_report_html}}
}
