% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{process_config_column}
\alias{process_config_column}
\title{Process Configuration Column}
\usage{
process_config_column(config_column, index, num_levels, mode)
}
\arguments{
\item{config_column}{A configuration column from the spline test
configurations.}

\item{index}{Index of the configuration to process.}

\item{num_levels}{Number of unique levels in the metadata condition.}

\item{mode}{A character string specifying the mode
('integrated' or 'isolated').}
}
\value{
A vector or list with the processed configuration values.
}
\description{
Processes a configuration column based on the given mode and number of
levels.
}
