% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{process_data}
\alias{process_data}
\title{utils scripts contains shared functions that are used by at least two package 
functions of the SplineOmics package.}
\usage{
process_data(data)
}
\description{
utils scripts contains shared functions that are used by at least two package 
functions of the SplineOmics package.
}
