% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{process_enrichment_results}
\alias{process_enrichment_results}
\title{Process Enrichment Results}
\usage{
process_enrichment_results(
  all_db_results,
  enrichment_results,
  adjP_threshold,
  column_name,
  count_column_name,
  background = FALSE
)
}
\arguments{
\item{all_db_results}{A list of data frames containing enrichment results
for all databases.}

\item{enrichment_results}{A list of data frames containing enrichment
results for individual databases.}

\item{adjP_threshold}{The threshold for adjusted p-values.}

\item{column_name}{The name of the column to store adjusted p-values.}

\item{count_column_name}{The name of the column to store gene counts.}

\item{background}{Logical indicating whether background ratios are included.}
}
\value{
A list of data frames containing processed enrichment results.
}
\description{
Process enrichment results for visualization.
}
