% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{process_field}
\alias{process_field}
\title{Process and Encode Data Field for Report}
\usage{
process_field(
  field,
  data,
  meta,
  topTables,
  report_info,
  encode_df_to_base64,
  report_type,
  enrichr_format
)
}
\arguments{
\item{field}{A string specifying the field to process.}

\item{data}{A dataframe containing the main data.}

\item{meta}{A dataframe containing meta information.}

\item{topTables}{A dataframe containing the results of differential 
expression analysis.}

\item{report_info}{A list containing additional report information.}

\item{encode_df_to_base64}{A function to encode a dataframe to base64.}

\item{report_type}{A string specifying the type of report.}

\item{enrichr_format}{A list with the formatted gene lists and background 
gene list.}
}
\value{
A string containing the HTML link for downloading the processed 
        field.
}
\description{
This function processes a given field, encodes the associated data as base64, 
and generates a download link for the report. It handles different types of 
fields including data, meta, top tables, and Enrichr formatted gene lists.
}
