% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{process_plots}
\alias{process_plots}
\title{Process Plots}
\usage{
process_plots(plots, plots_sizes, index, html_content)
}
\arguments{
\item{plots}{A list of plots to be processed.}

\item{plots_sizes}{A list of sizes for the plots.}

\item{index}{An integer, specifying which element index to take of plots and
plots_sizes.}

\item{html_content}{The current state of the HTML content.}
}
\value{
Updated HTML content with the plots included.
}
\description{
Converts plots to base64 and appends them to the HTML content.
}
