% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gsea_report.R
\name{process_result}
\alias{process_result}
\title{Process GSEA Result for a Specific Level}
\usage{
process_result(level_result, level_name)
}
\arguments{
\item{level_result}{A list containing the GSEA result for a specific level.}

\item{level_name}{A character string representing the name of the level.}
}
\value{
A list with the following components:
\describe{
  \item{plot}{A plot object or "section_break" if the result contains `NA`.}
  \item{plot_size}{An integer indicating the size of the plot.}
  \item{header_info}{A list with header information, including the level 
  name, full enrichment results, and raw enrichment results if available.}
}
}
\description{
This function processes the GSEA result for a specific level. It handles
cases where the result contains `NA` values by adding a section break.
Otherwise, it extracts the plot, plot size, and header information from 
the result.
}
