% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{process_top_table}
\alias{process_top_table}
\title{Converts top_table to a tibble and adds the columns feature_names and 
(Intercept) and appends it to the top_table list.}
\usage{
process_top_table(top_table_and_fit, feature_names)
}
\description{
Converts top_table to a tibble and adds the columns feature_names and 
(Intercept) and appends it to the top_table list.
}
