% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_limma_report.R
\name{remove_prefix}
\alias{remove_prefix}
\title{Remove Prefix from String}
\usage{
remove_prefix(string, prefix)
}
\arguments{
\item{string}{A string from which the prefix should be removed.}

\item{prefix}{A string representing the prefix to be removed.}
}
\value{
A string with the prefix removed.
}
\description{
Removes a specified prefix from the beginning of a string. This function
is useful for cleaning or standardizing strings by removing known prefixes.
}
