% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{run_gsea}
\alias{run_gsea}
\title{Generate a GSEA Report}
\usage{
run_gsea(
  levels_clustered_hits,
  genes,
  databases,
  report_info,
  params = NA,
  plot_titles = NA,
  background = NULL,
  report_dir = here::here()
)
}
\arguments{
\item{levels_clustered_hits}{A list of clustered hits at different levels.}

\item{genes}{A vector of genes to be analyzed.}

\item{databases}{A list of databases to be used in the analysis.}

\item{report_info}{A list containing information for the report generation.}

\item{params}{Additional parameters for the GSEA analysis, default is NA.}

\item{plot_titles}{Titles for the plots, default is NA.}

\item{background}{Background data, default is NULL.}

\item{report_dir}{Directory where the report will be saved, default is 
`here::here()`.}
}
\value{
A list of plots generated for the GSEA report.
}
\description{
This function generates a Gene Set Enrichment Analysis (GSEA) report based
on clustered hit levels, gene data, and specified databases. It processes
the input data, manages GSEA levels, and produces an HTML report with plots.
}
