% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{run_gsea_level}
\alias{run_gsea_level}
\title{Perform Gene Set Enrichment Analysis and plot it.}
\usage{
run_gsea_level(
  clustered_genes,
  databases,
  params = NA,
  plot_title = "",
  background = NULL
)
}
\arguments{
\item{clustered_genes}{A list of dataframes with two columns: the first
column contains the standard gene symbol, and the second column contains an
integer specifying the cluster.}

\item{databases}{A dataframe containing the data of the downloaded Enrichr
databases}

\item{params}{A list specifying the clusterProfiler parameters for the 
enrichment analysis.}

\item{plot_title}{An optional string specifying the title of the plot. If not
provided, a default title based on the analysis will be used.}

\item{background}{An optional list of standard gene symbols to be used as the
background for the enrichment analysis instead of the background chosen by
the `enricher`. The default is an empty list, which implies the use of the
default background set by the enrichment tool.}
}
\value{
An object containing the results of the Gene Set Enrichment Analysis,
including any plots generated during the analysis.
}
\description{
This function conducts a Gene Set Enrichment Analysis (GSEA) using either the
clusterProfiler package. Afterwards, it plots the results.
It allows for customization of enrichment parameters, selection of databases,
and optionally specifying a custom plot title and background gene list.
}
