% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{screen_limma_hyperparams}
\alias{screen_limma_hyperparams}
\title{screen_limma_hyperparams.R contains the exported package function 
screen_limma_hyperparams  and all the functions that make up the 
functionality of screen_limma_hyperparams. screen_limma_hyperparams runs the 
other package function, run_limma_splines, for a time series omics dataset,
for different hyperparameters. Such are for example degree of freedom of the
spline, type of spline, limma design formula, and different versions of the 
data (full data vs. outliers removed). This can result in several 
combinations, and it is tedious analyzing the combination in an organised 
and structured manner. Therefore, this function streamlines that process.
Limma Hyperparameters Screening}
\usage{
screen_limma_hyperparams(
  splineomics,
  datas,
  datas_descr,
  metas,
  designs,
  spline_test_configs,
  report_dir = here::here(),
  adj_pthresholds = c(0.05),
  time_unit = "min",
  padjust_method = "BH"
)
}
\arguments{
\item{datas}{A list of data frames containing the datasets to be analyzed.}

\item{datas_descr}{A description object for the data.}

\item{metas}{A list of data frames containing metadata for each dataset in 
`datas`.}

\item{designs}{A character vector of design formulas for the limma analysis.}

\item{spline_test_configs}{A configuration object for spline tests.}

\item{report_dir}{A non-empty string specifying the report directory.}

\item{adj_pthresholds}{A numeric vector of p-value thresholds for 
significance determination.}

\item{time_unit}{A character string specifying the time unit label for plots.}

\item{padjust_method}{A character string specifying the method for p-value 
adjustment.}

\item{condition}{A single character string specifying the condition.}

\item{report_info}{An object containing report information.}

\item{meta_batch_column}{A character string specifying the meta batch column.}

\item{meta_batch2_column}{A character string specifying the second meta 
batch column (the limma function removeBatchEffect
supports a maximum of two batch columns.)}
}
\value{
Returns a list of plots generated from the limma analysis results.
        Each element in the list corresponds to a different combination of 
        hyperparameters.
}
\description{
This function screens through various combinations of hyperparameters for 
limma analysis,
including designs, modes, and degrees of freedom. It validates inputs, 
generates results for all
combinations, and plots the outcomes. Finally, it may also be involved in 
generating an HTML report
as part of a larger analysis workflow.
}
