% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{set_default_params}
\alias{set_default_params}
\title{Set Default Parameters}
\usage{
set_default_params(params)
}
\arguments{
\item{params}{A list of parameters to be checked and updated with default
values if necessary.}
}
\value{
A list of parameters with all required elements, either from the
        input `params` or with added default values for any missing elements.
}
\description{
This function checks if the provided `params` list is `NA` or missing any
elements. If `params` is `NA`, it assigns a list of default parameters.
If any element is missing from `params`, it adds the missing element with
its respective default value.
}
