% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_limma_report.R
\name{shorten_names}
\alias{shorten_names}
\title{Shorten Names}
\usage{
shorten_names(name, unique_values)
}
\arguments{
\item{name}{A string representing the name to be shortened.}

\item{unique_values}{A vector of unique values whose abbreviations will
replace their occurrences in the name.}
}
\value{
A string with the unique values replaced by their abbreviations.
}
\description{
Replaces occurrences of unique values within a name with their first three
characters. This function is useful for abbreviating long condition names
in a dataset.
}
