% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{stop_call_false}
\alias{stop_call_false}
\title{Stop with custom message without call.}
\usage{
stop_call_false(...)
}
\arguments{
\item{...}{One or more character strings specifying the error message.
If multiple strings are provided, they will be concatenated
with a space between them.}
}
\value{
This function does not return a value; it stops execution and
        throws an error.
}
\description{
A helper function that triggers an error with the specified message and
suppresses the function call in the error output. This function behaves
similarly to the base `stop()` function but automatically concatenates
multiple message strings if provided.
}
