% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{store_hits}
\alias{store_hits}
\title{Store Hits}
\usage{
store_hits(condition)
}
\arguments{
\item{condition}{A list containing dataframes and parameters for each
condition.}
}
\value{
A list where each element is a vector of feature indices that meet
the significance threshold.
}
\description{
Stores the feature indices for significant hits based on the adjusted p-value
threshold for each condition.
}
