% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{truncate_row_names}
\alias{truncate_row_names}
\title{Truncate Row Names}
\usage{
truncate_row_names(names, max_length = 40)
}
\arguments{
\item{names}{A character vector of row names.}

\item{max_length}{An integer specifying the maximum length of the row names.
Default is 40.}
}
\value{
A character vector of truncated row names.
}
\description{
This function truncates row names that exceed a specified maximum length.
If the row name length exceeds the maximum length, it appends " ..."
to indicate truncation.
}
