% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineomics_object.R
\name{update_splineomics}
\alias{update_splineomics}
\title{Update the variables in a SplineOmics object}
\usage{
update_splineomics(splineomics, ...)
}
\arguments{
\item{splineomics}{\code{SplineOmics}: A SplineOmics object to be updated.}

\item{...}{Named arguments with new values for fields to be updated or added.}
}
\value{
The updated SplineOmics object.
}
\description{
Updates a SplineOmics object by modifying existing fields or adding new ones.
}
\examples{
set.seed(1)
toy_data <- matrix(rnorm(12),
    nrow = 3,
    dimnames = list(paste0("gene", 1:3), paste0("S", 1:4))
)
toy_meta <- data.frame(
    SampleID = colnames(toy_data),
    Time = c(0, 0, 1, 1),
    Condition = c("Ctrl", "Ctrl", "Trt", "Trt"),
    stringsAsFactors = FALSE,
    row.names = colnames(toy_data)
)

so <- create_splineomics(
    data = toy_data,
    meta = toy_meta,
    condition = toy_meta$Condition
)

# Update the mode and design formula
new_design <- "~ 1 + Condition*Time"
so_updated <- update_splineomics(so,
    mode = "integrated",
    design = new_design
)

str(so_updated, max.level = 1)

}
