% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{between_level}
\alias{between_level}
\title{Performs limma spline analysis and gets the top_table for the level 
comparison: Both for factor and for factor:time}
\usage{
between_level(
  data,
  meta,
  design,
  DoFs,
  condition,
  compared_levels,
  padjust_method,
  feature_names
)
}
\description{
Performs limma spline analysis and gets the top_table for the level 
comparison: Both for factor and for factor:time
}
