% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{between_level}
\alias{between_level}
\title{Between Level Analysis}
\usage{
between_level(
  data,
  preprocess_rna_seq,
  meta,
  design,
  spline_params,
  condition,
  compared_levels,
  padjust_method,
  feature_names
)
}
\arguments{
\item{data}{A matrix of data values.}

\item{preprocess_rna_seq}{Boolean specifying whether to preprocess RNA seq}

\item{meta}{A dataframe containing metadata, including a 'Time' column.}

\item{design}{A design formula or matrix for the LIMMA analysis.}

\item{spline_params}{A list of spline parameters for the analysis.}

\item{condition}{A character string specifying the condition.}

\item{compared_levels}{A vector of levels within the condition to compare.}

\item{padjust_method}{A character string specifying the p-adjustment method.}

\item{feature_names}{A non-empty character vector of feature names.}
}
\value{
A list containing top tables for the factor only and factor-time 
contrast.
}
\description{
Performs a between-level analysis using LIMMA to compare specified levels 
within a condition.
}
\seealso{
\code{\link[splines]{bs}}, \code{\link[splines]{ns}}, 
\code{\link[limma]{lmFit}}, \code{\link[limma]{eBayes}}, 
\code{\link[limma]{topTable}}, \code{\link{modify_limma_top_table}}
}
