% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{check_dataframe}
\alias{check_dataframe}
\title{Check Dataframe}
\usage{
check_dataframe(df)
}
\arguments{
\item{df}{A dataframe to check.}
}
\value{
TRUE if the dataframe is valid, otherwise an error is thrown.
}
\description{
Validates that the dataframe contains all required columns with the correct
data types.
}
