% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{check_design_formula}
\alias{check_design_formula}
\title{Check Design Formula}
\usage{
check_design_formula(formula, meta)
}
\arguments{
\item{formula}{A character string representing the design formula.}

\item{meta}{A dataframe containing metadata.}
}
\value{
TRUE if the design formula is valid, otherwise an error is thrown.
}
\description{
Validates the design formula ensuring it is a valid character string, 
contains allowed characters, includes the intercept term 'X', and references 
columns present in the metadata.
}
\examples{
meta <- data.frame(Time = seq(1, 10), condition = rep(c("A", "B"), each = 5))
check_design_formula("~ Time + condition * X", meta)

}
\seealso{
\code{\link{stats::model.matrix}}
}
