% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{check_max_and_min_dof}
\alias{check_max_and_min_dof}
\title{Calculate degrees of freedom for spline configurations}
\usage{
check_max_and_min_dof(spline_test_configs, metas)
}
\arguments{
\item{spline_test_configs}{A dataframe with columns including 'spline_type',
'dof', 'degree', and 'knots'.}

\item{meta}{A dataframe with a 'Time' column used to validate the maximum 
allowable DoF.}
}
\value{
NULL The function stops if any DoF condition is not met, with an 
error message.
}
\description{
This function calculates and checks degrees of freedom for each spline 
configuration
based on the spline type. It validates DoF against constraints given in the 
meta dataframe.
}
