% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{check_max_and_min_dof}
\alias{check_max_and_min_dof}
\title{Check Maximum and Minimum Degrees of Freedom}
\usage{
check_max_and_min_dof(spline_test_configs, metas)
}
\arguments{
\item{spline_test_configs}{A dataframe containing spline test configurations.}

\item{metas}{A list of metadata corresponding to the data matrices.}
}
\value{
No return value, called for side effects.
}
\description{
Validates the degrees of freedom (DoF) for each row in the spline test 
configurations based on the metadata.
}
