% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_meta}
\alias{check_meta}
\title{Check Metadata}
\usage{
check_meta(meta, condition, meta_batch_column = NA, data_meta_index = NA)
}
\arguments{
\item{meta}{A dataframe containing the metadata, including the 'Time' column.}

\item{condition}{A single character string specifying the column name in the 
meta dataframe to be checked.}

\item{meta_batch_column}{An optional parameter specifying the column name in
the meta dataframe used to remove the batch effect. Default is NA.}

\item{data_meta_index}{An optional parameter specifying the index of the 
data/meta pair for error messages. Default is NA.}
}
\value{
Returns TRUE if all checks pass. Stops execution and returns an 
error message if any check fails.
}
\description{
This function checks the validity of the metadata dataframe, ensuring it 
contains the 'Time' column, 
does not contain missing values, and that the specified condition column is 
valid and of the appropriate type.
Additionally, it checks for an optional batch effect column and prints 
messages regarding its use.
}
\examples{
\dontrun{
# Example of how to use the function
meta <- data.frame(Time = 1:3, Condition = c("A", "B", "C"))
check_meta(meta, "Condition")

invalid_meta <- data.frame(Time = c(1, NA, 3), Condition = c("A", "B", "C"))
try(check_meta(invalid_meta, "Condition"))  # Should raise an error}

}
