% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{check_mode}
\alias{check_mode}
\title{Check Mode}
\usage{
check_mode(mode)
}
\arguments{
\item{mode}{A character string specifying the mode.}
}
\value{
A message indicating the chosen mode if valid; otherwise, an error 
is thrown.
}
\description{
Validates that the mode is either 'integrated' or 'isolated', which depends 
on the design formula used in limma.
}
\examples{
\dontrun{
check_mode("integrated")
check_mode("isolated")}

}
\seealso{
\code{\link{limma}}
}
