% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_spline_params}
\alias{check_spline_params}
\title{Check Spline Parameters}
\usage{
check_spline_params(spline_params, mode)
}
\arguments{
\item{spline_params}{A list of spline parameters.}

\item{mode}{A character string specifying the mode
('integrated' or 'isolated').}
}
\value{
No return value, called for side effects.
}
\description{
Validates the spline parameters both generally and depending on the 
specified mode.
}
\examples{
spline_params <- list(spline_type = c("n"), dof = list(3))
mode <- "integrated"
check_spline_params(spline_params, mode)

}
\seealso{
\code{\link{check_spline_params_generally}}, 
\code{\link{check_spline_params_mode_dependent}}
}
