% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{check_spline_type_params}
\alias{check_spline_type_params}
\title{Check Spline Type Parameters}
\usage{
check_spline_type_params(spline_test_configs)
}
\arguments{
\item{spline_test_configs}{A dataframe containing spline test configurations.}
}
\value{
TRUE if all checks pass, otherwise an error is thrown.
}
\description{
Validates the parameters for each row in the spline test configurations 
based on the spline type.
}
