% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{check_spline_type_params}
\alias{check_spline_type_params}
\title{Check constraints on spline configuration based on spline_type}
\usage{
check_spline_type_params(spline_test_configs)
}
\arguments{
\item{spline_test_configs}{A dataframe with columns 'spline_type', 'degree',
'dof', 'knots', 'bknots'.}
}
\value{
logical TRUE if all conditions are met, otherwise stops with an error.
}
\description{
This function verifies the conditions in the spline_test_configs dataframe.
It checks the 'spline_type' column to dictate rules for 'degree', 'dof',
'knots', and 'bknots'.
}
