% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{cluster_hits}
\alias{cluster_hits}
\title{Cluster Hits from Top Tables}
\usage{
cluster_hits(top_tables, data, meta, condition, p_values, clusters, report_dir)
}
\arguments{
\item{top_tables}{A list of data frames, each representing a top table from 
differential expression
       analysis, containing at least 'adj.P.Val' and expression data columns.}

\item{data}{The original expression dataset used for differential expression 
analysis.}

\item{meta}{Metadata dataframe corresponding to the `data`, must include a 
'Time' column and any columns
       specified by `conditions`.}

\item{p_values}{Numeric vector of p-value thresholds for filtering hits in 
each top table.}

\item{clusters}{Integer vector specifying the number of clusters to cut the 
dendrogram into, for each
       group factor.}

\item{report_dir}{Character string specifying the directory path where the 
HTML report and any
       other output files should be saved.}

\item{conditions}{Character vector specifying the column names in `meta` used 
to define groups for
       analysis.}
}
\value{
A list where each element corresponds to a group factor and contains 
the clustering results,
        including `clustered_hits` data frame, hierarchical clustering object 
        `hc`, `curve_values`
        data frame with normalized spline curves, and `top_table` with 
        cluster assignments.
}
\description{
Performs clustering on hits from top tables generated by differential 
expression analysis.
This function filters hits based on adjusted p-value thresholds, extracts 
spline coefficients for 
significant features, normalizes these coefficients, and applies hierarchical 
clustering. The results,
including clustering assignments and normalized spline curves, are saved in a 
specified directory and
compiled into an HTML report.
}
\examples{
\dontrun{
  cluster_results <- cluster_hits(top_tables, data, meta, c("GroupFactor"), 
                                  c(0.05),
                                  c(3), "path/to/report/dir")
}


}
\seealso{
\code{\link[limma]{topTable}}, \code{\link[stats]{hclust}}
}
