% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{cluster_hits}
\alias{cluster_hits}
\title{cluster_hits.R contains the exported package function cluster_hits and all
the functions that make up the functionality of cluster_hits. cluster_hits
clusters the hits of a time series omics datasets (the features that were
significantly changed over the time course) with hierarchical clustering of
the spline shape.
Cluster Hits from Top Tables}
\usage{
cluster_hits(
  top_tables,
  data,
  meta,
  design,
  condition,
  report_info,
  spline_params = list(spline_type = c("n"), dof = c(2L)),
  adj_pthresholds = c(0.05),
  clusters = c("auto"),
  meta_batch_column = NA,
  meta_batch2_column = NA,
  time_unit = "min",
  report_dir = here::here(),
  report = TRUE
)
}
\arguments{
\item{top_tables}{A list of data frames, each representing a top table from
differential expression analysis, containing at least 'adj.P.Val' and
expression data columns.}

\item{data}{The original expression dataset used for differential expression
analysis.}

\item{meta}{A dataframe containing metadata corresponding to the `data`,
must include a 'Time' column and any columns specified by `conditions`.}

\item{design}{A character of length 1 representing the limma design formula}

\item{condition}{Character of length 1 specifying the column name in `meta`
used to define groups for analysis.}

\item{report_info}{A character string to be printed at the top of the report.}

\item{spline_params}{A list of spline parameters for the analysis.}

\item{adj_pthresholds}{Numeric vector of p-value thresholds for filtering
hits in each top table.}

\item{clusters}{Character or integer vector specifying the number of
clusters or 'auto' for automatic estimation.}

\item{meta_batch_column}{A character string specifying the column name in
the metadata used for batch effect removal.}

\item{time_unit}{A character string specifying the time unit label for plots
(e.g., 'm' for minutes).}

\item{report_dir}{Character string specifying the directory path where the
HTML report and any other output files should be saved.}

\item{report}{Boolean TRUE or FALSE value specifing if a report should be
generated.}

\item{mode}{A character string specifying the mode ('integrated' or
'isolated').}
}
\value{
A list where each element corresponds to a group factor and contains
the clustering results,
        including `clustered_hits` data frame, hierarchical clustering object
        `hc`, `curve_values`
        data frame with normalized spline curves, and `top_table` with
        cluster assignments.
}
\description{
Performs clustering on hits from top tables generated by differential
expression analysis.
This function filters hits based on adjusted p-value thresholds, extracts
spline coefficients for
significant features, normalizes these coefficients, and applies hierarchical
clustering. The results,
including clustering assignments and normalized spline curves, are saved in a
specified directory and
compiled into an HTML report.
}
\examples{
\dontrun{
  cluster_results <- cluster_hits(top_tables, data, meta, c("GroupFactor"),
                                  c(0.05),
                                  c(3), "path/to/report/dir")
}

}
\seealso{
\code{\link[limma]{topTable}}, \code{\link[stats]{hclust}}
}
