% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{control_inputs_run_limma}
\alias{control_inputs_run_limma}
\title{Controls the function input and raises an error if not in expected format}
\usage{
control_inputs_run_limma(
  data,
  meta,
  design,
  DoFs,
  condition,
  feature_names,
  mode,
  padjust_method
)
}
\description{
Controls the function input and raises an error if not in expected format
}
