% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_limma_report.R
\name{create_limma_report}
\alias{create_limma_report}
\title{Create a limma report}
\usage{
create_limma_report(splineomics, adj_pthresh = 0.05, report_dir = here::here())
}
\arguments{
\item{adj_pthresh}{A numeric value specifying the adjusted p-value threshold 
for significance. Default is 0.05.}

\item{report_dir}{A string specifying the directory where the report should 
be saved. Default is the current working directory.}

\item{run_limma_splines_result}{A list containing the results of the LIMMA 
analysis with splines. It should have three components: `time_effect`, 
`avrg_diff_conditions`, and `interaction_condition_time`.}

\item{report_info}{A list containing metadata and other information to be 
included in the report.}
}
\value{
A list of plots included in the generated HTML report.
}
\description{
Generates an HTML report based on the results of a limma analysis with 
splines. 
The report includes various plots and sections summarizing the analysis 
results for time effects, average differences between conditions, 
and interaction effects between condition and time.
}
