% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineomics_object.R
\name{create_splineomics}
\alias{create_splineomics}
\title{Create and update the SplineOmics object
=====}
\usage{
create_splineomics(
  data,
  meta,
  condition,
  annotation = NULL,
  report_info = NULL,
  meta_batch_column = NULL,
  meta_batch2_column = NULL,
  feature_name_columns = NULL,
  design = NULL,
  spline_params = NULL,
  preprocess_rna_seq = FALSE,
  normalization_fun = NULL,
  padjust_method = "BH"
)
}
\arguments{
\item{data}{The actual omics data.}

\item{meta}{Metadata associated with the omics data.}

\item{condition}{A condition variable.}

\item{annotation}{A dataframe with the feature descriptions of data 
(optional).}

\item{report_info}{A list containing report information such as omics data 
type, data description, data collection date, analyst name, contact info, 
and project name (optional).}

\item{meta_batch_column}{Column for meta batch information (optional).}

\item{meta_batch2_column}{Column for secondary meta batch information 
(optional).}

\item{feature_name_columns}{Character vector containing the column names of 
the annotation info that describe the features.
This argument is used to specify in the HTML 
report how exactly the feature names displayed
above each individual spline plot have been
created. Use the same vector that was used to 
create the row headers for the data matrix!}

\item{design}{A design matrix or similar object (optional).}

\item{spline_params}{Parameters for spline functions (optional).}

\item{preprocess_rna_seq}{Boolean specifying whether to preprocess RNA seq}

\item{normalization_fun}{Function used for normalizing RNA-seq. Must take as
input the y of: y <- edgeR::DGEList(counts = raw_counts) and output the y 
with the normalized counts.}

\item{padjust_method}{Method for p-value adjustment, one of "none", "BH", 
"BY", "holm", "bonferroni", "hochberg", or "hommel". 
Defaults to "BH" (Benjamini-Hochberg).}
}
\value{
A SplineOmics object.
}
\description{
Creates a SplineOmics object containing variables that are commonly used 
across multiple functions in the package.
}
\details{
Description
-----------
Contains the functions to create and update a SplineOmics object. This object
is used to collect function arguments, that are equivalent for more than one 
exported function of the SplineOmics package. Additionally

Functions
---------
- create_splineomics: Create a SplineOmics object
- update_splineomics: Add additional arguments to the SplineOmics 
object or overwrite existing arguments.

Classes
-------
None

Notes
-----
None
Create a SplineOmics object
}
