% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineomics_object.R
\name{create_splineomics}
\alias{create_splineomics}
\title{Create and update the SplineOmics object
=====}
\usage{
create_splineomics(
  data,
  meta,
  condition,
  annotation = NULL,
  report_info = NULL,
  meta_batch_column = NULL,
  meta_batch2_column = NULL,
  feature_name_columns = NULL,
  design = NULL,
  spline_params = NULL,
  preprocess_rna_seq = FALSE,
  padjust_method = "BH"
)
}
\arguments{
\item{data}{The actual omics data.}

\item{meta}{Metadata associated with the omics data.}

\item{condition}{A condition variable.}

\item{annotation}{A dataframe with the feature descriptions of data 
(optional).}

\item{report_info}{A list containing report information such as omics data 
type, data description, data collection date, analyst name, contact info, 
and project name (optional).}

\item{meta_batch_column}{Column for meta batch information (optional).}

\item{meta_batch2_column}{Column for secondary meta batch information 
(optional).}

\item{design}{A design matrix or similar object (optional).}

\item{spline_params}{Parameters for spline functions (optional).}
}
\value{
A SplineOmics object.
}
\description{
Creates a SplineOmics object containing variables that are commonly used 
across multiple functions in the package.
}
\details{
Description
-----------
Contains the functions to create and update a SplineOmics object. This object
is used to collect function arguments, that are equivalent for more than one 
exported function of the SplineOmics package. Additionally

Functions
---------
- create_splineomics: Create a SplineOmics object
- update_splineomics: Add additional arguments to the SplineOmics 
object or overwrite existing arguments.

Classes
-------
None

Notes
-----
None
Create a SplineOmics object
}
