% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_enrichr_databases.R
\name{download_enrichr_databases}
\alias{download_enrichr_databases}
\title{Download the Enrichr databases}
\usage{
download_enrichr_databases(
  gene_set_lib,
  output_dir = here::here(),
  filename = NULL
)
}
\arguments{
\item{gene_set_lib}{\code{character()}: A character vector of database names to
download from Enrichr, for example:
\code{c("WikiPathways_2019_Human", "NCI-Nature_2016")}.}

\item{output_dir}{\code{character(1)}: A string specifying the output directory
where the \code{.tsv} file will be saved. Defaults to the current project
directory as defined by \code{here::here()}.}

\item{filename}{\code{character(1)}: Name of the output file (with file
extension). Due to commas present in some terms, \code{.tsv} is recommended.
When left out, the file is named \code{all_databases_timestamp.tsv}.}
}
\value{
A \code{data.frame} of gene set annotations with three columns:
\describe{
\item{DB}{Database name (e.g. \code{"WikiPathways_2019_Human"},
\code{"NCI-Nature_2016"}).}
\item{Geneset}{The gene set or pathway term from that database.}
\item{Gene}{A gene contained in the gene set.}
}

In addition to returning the \code{data.frame}, the function also writes the same
table to disk as a \code{.tsv} file in the specified \code{output_dir}.
}
\description{
This function downloads gene sets from specified Enrichr databases and saves
them to a specified output directory as a .tsv file per default. The file is
named with a timestamp per default to ensure uniqueness (all databases are
stored in a single file). This file has 3 columns: DB containing the database
name, Geneset, with the genesets, and Gene, with the gene names.
}
\examples{
if (interactive()) {
  libs <- c("WikiPathways_2019_Human")
  out <- download_enrichr_databases(
    gene_set_lib = libs,
    output_dir = tempdir(),
    filename = "enrichr_demo.tsv"
  )
  head(out)
}

}
