% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{explore_data}
\alias{explore_data}
\title{Generate Exploratory Plots}
\usage{
explore_data(
  data,
  meta,
  condition,
  report_info,
  meta_batch_column = NA,
  report_dir = here::here()
)
}
\arguments{
\item{data}{A numeric matrix containing the data.}
}
\value{
A list of ggplot objects representing various exploratory plots.
}
\description{
This function takes a data matrix, checks its validity, and generates a list 
of exploratory plots including density plots, boxplots, PCA plots, MDS plots, 
variance explained plots, and violin plots.
}
\examples{
\dontrun{
# Example usage
data <- matrix(rnorm(1000), ncol = 10)
plots <- generate_plots(data)
for (plot in plots) {
  print(plot)
}}

}
