% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{extract_data}
\alias{extract_data}
\title{utils scripts contains shared functions that are used by at least two package 
functions of the SplineOmics package.
Extract Numeric Matrix from Dataframe}
\usage{
extract_data(data)
}
\arguments{
\item{data}{A dataframe loaded from a csv file, potentially containing a 
rectangular or quadratic area with numeric data amidst other values.}
}
\value{
A matrix containing only the numeric data from the identified area, 
with rows containing NA values removed.
}
\description{
This function takes a dataframe and identifies a rectangular or quadratic 
area containing numeric data, starting from the first occurrence of a 
6x6 block of numeric values. It then extracts this area into a matrix, 
ensuring that each row contains only numeric values. Rows with any NA values 
are removed from the resulting matrix.
}
\examples{
\dontrun{
# Example of how to use the function
data <- read.csv("path_to_your_file.csv")
result_matrix <- extract_numeric_matrix(data)
print(result_matrix)}

}
