% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{flatten_spline_configs}
\alias{flatten_spline_configs}
\title{Flatten Spline Configurations}
\usage{
flatten_spline_configs(spline_configs)
}
\arguments{
\item{spline_configs}{A list of spline configuration objects.}
}
\value{
A list of formatted strings representing each spline configuration.
}
\description{
Flattens and formats spline configurations into a list of formatted strings.
}
