% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{gen_composite_spline_plots}
\alias{gen_composite_spline_plots}
\title{Generate Composite Spline Plots}
\usage{
gen_composite_spline_plots(
  internal_combos,
  datas,
  metas,
  spline_test_configs,
  time_unit_label
)
}
\arguments{
\item{internal_combos}{A list containing combinations of top tables.}

\item{datas}{A list of matrices.}

\item{metas}{A list of metadata corresponding to the data matrices.}

\item{spline_test_configs}{A configuration object for spline tests.}

\item{time_unit_label}{A character string specifying the time unit label 
for plots.}
}
\value{
A list containing the composite spline plots and their lengths.
}
\description{
Creates composite spline plots for significant and non-significant features 
across multiple levels within a condition.
One half of one condition comparison HTML 
(composite spline plots for one 'condition' inside one condition comparison)
}
\seealso{
\code{\link{plot_composite_splines}}
}
