% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{gen_composite_spline_plots}
\alias{gen_composite_spline_plots}
\title{One half of one condition comparison HTML 
(composite spline plots for one 'condition' inside one condition comparison)}
\usage{
gen_composite_spline_plots(internal_combos, datas, metas, spline_test_configs)
}
\description{
One half of one condition comparison HTML 
(composite spline plots for one 'condition' inside one condition comparison)
}
