% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{generate_and_write_html}
\alias{generate_and_write_html}
\title{Generate and Write HTML Report}
\usage{
generate_and_write_html(toc, html_content, report_info, output_file_path)
}
\arguments{
\item{toc}{A string containing the table of contents in HTML format.}

\item{html_content}{A string containing the main HTML content with a
placeholder for the table of contents.}

\item{report_info}{A list containing report information such as
`contact_info` and `analyst_name`.}

\item{output_file_path}{A string specifying the path where the final
HTML file will be written.}
}
\description{
This function generates an HTML report by inserting a table of contents,
embedding necessary JavaScript files, and writing the final HTML content
to a specified output file.
}
