% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{generate_explore_plots}
\alias{generate_explore_plots}
\title{#' Control Input Parameters for Data Exploration Functions
#'
#' @description
#' This function validates the input parameters for data exploration functions.
#' It checks the structure and data types of `data`, `meta`, `condition`,
#' `report_info`, `meta_batch_column`, `meta_batch2_column`, `report_dir`, 
#' and `report`. If any condition is not met, the function stops the script 
#' and produces an informative error message.
#'
#' @param data A dataframe containing the data to be explored.
#' @param meta A dataframe containing metadata associated with the data.
#' @param condition A character string specifying the condition to be explored.
#' @param report_info A list containing information for generating the report.
#' @param meta_batch_column A character string specifying the batch column in 
#'        the metadata.
#' @param meta_batch2_column A character string specifying the second batch 
#'        column in the metadata.
#' @param report_dir A character string specifying the directory where the 
#'        report will be saved.
#' @param report A Boolean indicating whether to generate a report (TRUE) or 
#'        not (FALSE).
#'
#' @return This function does not return a value. It stops with an error message 
#'         if the conditions are not met.
#'
control_inputs_explore_data <- function(data,
                                        meta,
                                        condition,
                                        report_info,
                                        meta_batch_column,
                                        meta_batch2_column,
                                        report_dir,
                                        report) {
  
  check_data_and_meta(data, 
                      meta, 
                      condition, 
                      meta_batch_column,
                      meta_batch2_column)
  
  check_report_info(report_info)
  
  check_and_create_report_dir(report_dir)
  
  
  if (report != TRUE && report != FALSE) {
    stop("report must be either Boolean TRUE or FALSE", call. = FALSE)
  }
}
Generate exploratory plots}
\usage{
generate_explore_plots(data, meta, condition)
}
\arguments{
\item{data}{A data frame or matrix containing the data to be plotted.}

\item{meta}{A data frame containing metadata associated with the data.}

\item{condition}{A string specifying the column in the metadata that contains 
the condition or grouping variable.}
}
\value{
A list containing two elements:
\describe{
  \item{plots}{A list of ggplot objects representing the generated plots.}
  \item{plots_sizes}{A vector of numeric values indicating the sizes of the 
  corresponding plots.}
}
}
\description{
This function generates various exploratory plots including density plots, 
box plots, violin plots, PCA plots, and correlation heatmaps based on the 
provided data and metadata.
}
