% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{generate_report_html}
\alias{generate_report_html}
\title{utils scripts contains shared functions that are used by at least two package 
functions of the SplineOmics package. The level separation is only valid
internally in this script, and has no connection to the script level of the
respective exported functions scripts.
Generate Report HTML}
\usage{
generate_report_html(
  plots,
  plots_sizes,
  report_info,
  data = NULL,
  meta = NA,
  topTables = NA,
  enrichr_format = NA,
  level_headers_info = NA,
  spline_params = NA,
  report_type = "explore_data",
  analysis_type = NA,
  mode = NA,
  filename = "report",
  timestamp = format(Sys.time(), "\%d_\%m_\%Y-\%H_\%M_\%S"),
  report_dir = here::here()
)
}
\arguments{
\item{plots}{A list of ggplot2 plot objects.}

\item{plots_sizes}{A list of integers specifying the size of each plot.}

\item{report_info}{A named list containing report information.}

\item{data}{A dataframe or a list of dataframes, containing data that should
be directly embedded in the HTML report for downloading.}

\item{meta}{A dataframe, containing metadata that should
be directly embedded in the HTML report for downloading.}

\item{level_headers_info}{A list of header information for each level.}

\item{spline_params}{A list of spline parameters.}

\item{report_type}{A character string specifying the report type 
('screen_limma_hyperparams' or 'cluster_hits').}

\item{mode}{A character string specifying the mode 
('isolated' or 'integrated').}

\item{filename}{A character string specifying the filename for the report.}

\item{timestamp}{A timestamp to include in the report filename.}

\item{report_dir}{A character string specifying the report directory.}
}
\value{
No return value, called for side effects.
}
\description{
Generates an HTML report with the provided plots, spline parameters, and 
report information.
}
\seealso{
\code{\link{build_hyperparams_screen_report}}, 
\code{\link{build_cluster_hits_report}}
}
