% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{generate_reports}
\alias{generate_reports}
\title{Generate Reports}
\usage{
generate_reports(combo_pair_plots, report_info, report_dir, timestamp)
}
\arguments{
\item{combo_pair_plots}{A list of plots for each pair of combinations.}

\item{report_info}{An object containing report information.}

\item{report_dir}{A non-empty string specifying the report directory.}

\item{timestamp}{A timestamp to include in the reports.}
}
\value{
No return value, called for side effects.
}
\description{
Builds HTML reports for all pairwise hyperparameter combination comparisons.
}
\examples{
generate_reports(combo_pair_plots = list(plot1, plot2),
                 report_info = "example_info", 
                 report_dir = "example_dir", 
                 timestamp = Sys.time())
                 
}
