% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{get_limma_combos_results}
\alias{get_limma_combos_results}
\title{Generate LIMMA Combination Results}
\usage{
get_limma_combos_results(
  datas,
  metas,
  designs,
  modes,
  condition,
  spline_test_configs,
  feature_names,
  adj_pthresh,
  padjust_method
)
}
\arguments{
\item{datas}{A list of matrices.}

\item{metas}{A list of metadata corresponding to the data matrices.}

\item{designs}{A list of design matrices.}

\item{modes}{A character vector containing 'isolated' or 'integrated'.}

\item{condition}{A single character string specifying the condition.}

\item{spline_test_configs}{A configuration object for spline tests.}

\item{feature_names}{A character vector of feature names.}

\item{adj_pthresh}{A numeric vector with elements > 0 and < 1.}

\item{padjust_method}{A single character string specifying the p-adjustment 
method.}
}
\value{
A list of results for each combination of data, design, and spline 
configuration.
}
\description{
Computes results for various combinations of data, design matrices, and 
spline configurations using the LIMMA method.
}
\examples{
get_limma_combos_results(datas = list(matrix(1:4, 2, 2)), 
                         metas = list(data.frame(batch = 1:2)), 
                         designs = list(matrix(1:4, 2, 2)), 
                         modes = c("isolated"), 
                         condition = "example_condition", 
                         spline_test_configs = 
                            list(spline_type = c("type1")), 
                         feature_names = c("feature1", "feature2"), 
                         adj_pthresh = c(0.05), 
                         padjust_method = "BH")
                         
}
