% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{hc_barplot}
\alias{hc_barplot}
\title{Generate Barplot for Hit Comparison Object}
\usage{
hc_barplot(hc_obj)
}
\arguments{
\item{hc_obj}{An object of class "hitcomp" containing hit data for two 
conditions.}
}
\value{
A ggplot2 object representing the barplot.
}
\description{
Creates a barplot to visualize the number of significant features for each 
parameter set in the hit comparison object.
}
\seealso{
\code{\link{store_hits}}, \code{\link{ggplot2}}
}
