% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{is_not_na}
\alias{is_not_na}
\title{Check if Not All Values are NA}
\usage{
is_not_na(x)
}
\arguments{
\item{x}{An atomic vector or any other object.}
}
\value{
TRUE if the vector contains at least one non-NA value or if the 
object is not atomic; FALSE otherwise.
}
\description{
Determines if a given atomic vector contains at least one non-NA value.
}
\examples{
is_not_na(c(NA, NA, 3)) # returns TRUE
is_not_na(c(NA, NA, NA)) # returns FALSE
is_not_na(list(NA, 1, NA)) # returns TRUE

}
