% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_validation.R
\name{Level3Functions}
\alias{Level3Functions}
\title{Level3Functions: A class for level 3 utility functions}
\description{
Level3Functions: A class for level 3 utility functions

Level3Functions: A class for level 3 utility functions
}
\details{
This class provides methods for creating error messages and checking
batch columns.

Check Batch Column
}
\section{Super class}{
\code{\link[SplineOmics:Level4Functions]{SplineOmics::Level4Functions}} -> \code{Level3Functions}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Level3Functions-check_batch_column}{\code{Level3Functions$check_batch_column()}}
\item \href{#method-Level3Functions-clone}{\code{Level3Functions$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SplineOmics" data-topic="Level4Functions" data-id="create_error_message"><a href='../../SplineOmics/html/Level4Functions.html#method-Level4Functions-create_error_message'><code>SplineOmics::Level4Functions$create_error_message()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Level3Functions-check_batch_column"></a>}}
\if{latex}{\out{\hypertarget{method-Level3Functions-check_batch_column}{}}}
\subsection{Method \code{check_batch_column()}}{
This method checks the batch column in the metadata and provides 
appropriate messages.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Level3Functions$check_batch_column(meta, meta_batch_column, data_meta_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meta}}{A dataframe containing metadata.}

\item{\code{meta_batch_column}}{A character string specifying the batch column
in the metadata.}

\item{\code{data_meta_index}}{An optional parameter specifying the index of the
data/meta pair. Default is NA.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL. The method is used for its side effects of throwing errors 
or printing messages.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Level3Functions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Level3Functions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Level3Functions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
