% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{make_clustering_report}
\alias{make_clustering_report}
\title{Make Clustering Report}
\usage{
make_clustering_report(
  all_levels_clustering,
  condition,
  data,
  meta,
  spline_params,
  adj_pthresholds,
  report_dir,
  mode,
  report_info,
  design,
  meta_batch_column,
  time_unit = time_unit
)
}
\arguments{
\item{all_levels_clustering}{A list containing clustering results for each 
level within a condition.}

\item{condition}{A character string specifying the condition.}

\item{data}{A matrix of data values.}

\item{meta}{A dataframe containing metadata.}

\item{spline_params}{A list of spline parameters for the analysis.}

\item{adj_pthresholds}{A numeric vector of p-values.}

\item{report_dir}{A character string specifying the report directory.}

\item{mode}{A character string specifying the mode 
('isolated' or 'integrated').}

\item{report_info}{An object containing report information.}

\item{design}{A string representing the limma design formula}

\item{meta_batch_column}{A character string specifying the meta batch column.}

\item{time_unit}{A character string specifying the time unit 
('s', 'm', 'h', 'd').}
}
\value{
No return value, called for side effects.
}
\description{
Generates a detailed clustering report including heatmaps, dendrograms, 
curve plots, and consensus shapes for each level within a condition.
}
\seealso{
\code{\link{removeBatchEffect}}, \code{\link{plot_heatmap}}, 
\code{\link{plot_dendrogram}}, \code{\link{plot_all_shapes}}, 
\code{\link{plot_consensus_shapes}}, \code{\link{plot_splines}}, 
\code{\link{generate_report_html}}
}
