% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_density_plots}
\alias{make_density_plots}
\title{Generate Density Plot}
\usage{
make_density_plots(data, meta, condition)
}
\arguments{
\item{data}{A numeric matrix containing the data.}
}
\value{
A ggplot object representing the density plot.
}
\description{
This function generates a density plot for a given data matrix. The density
plot shows the distribution of the values in the data matrix.
}
\examples{
\dontrun{
# Example usage
data <- matrix(rnorm(1000), ncol = 10)
density_plot <- make_density_plot(data)
print(density_plot)}

}
