% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{make_mds_plot}
\alias{make_mds_plot}
\title{Generate MDS Plot}
\usage{
make_mds_plot(data, meta, condition)
}
\arguments{
\item{data}{A numeric matrix containing the data.}
}
\value{
A ggplot object representing the MDS plot.
}
\description{
This function generates a multidimensional scaling (MDS) plot for a given 
data matrix. The MDS plot visualizes the similarities or dissimilarities 
between samples in the data matrix.
}
\examples{
# Example usage
data <- matrix(rnorm(1000), ncol = 10)
mds_plot <- make_mds_plot(data)
print(mds_plot)

}
