% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_data.R
\name{make_scatter_plot_html}
\alias{make_scatter_plot_html}
\title{make_scatter_plot_html()}
\usage{
make_scatter_plot_html(
  data,
  meta,
  output_file = "scatter_plots_base64.html",
  meta_replicate_column = NULL
)
}
\arguments{
\item{data}{A matrix with features as rows and samples as columns. 
Row names should be feature names.}

\item{meta}{A data frame with the meta information. Must contain a numeric
column "Time".}

\item{output_file}{The name of the HTML output file.}
}
\description{
This function is used to make scatter plots for the raw data of all the 
features. It generates an HTML report in the fashion of the other functions
of the SplineOmics package which contains all the scatter plots.
}
