% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gsea_report.R
\name{manage_gsea_level}
\alias{manage_gsea_level}
\title{Manage GSEA Analysis for a Specific Level}
\usage{
manage_gsea_level(clustered_hits, level_name)
}
\arguments{
\item{clustered_hits}{A dataframe containing the clustered hits for a 
specific level. It must include a column named `feature` to extract genes.}

\item{level_name}{A character string representing the name of the level.}
}
\value{
The result of the `create_gsea_report` function, which typically includes 
various plots and enrichment results.
}
\description{
This function manages the GSEA analysis for a specific level. It extracts
genes associated with the clustered hits, removes rows with `NA` values,
and runs the GSEA analysis using the `create_gsea_report` function.
}
