% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{modify_limma_top_table}
\alias{modify_limma_top_table}
\title{Modify limma Top Table}
\usage{
modify_limma_top_table(top_table, feature_names)
}
\arguments{
\item{top_table}{A dataframe containing the top table results from limma}

\item{feature_names}{A character vector of feature names.}
}
\value{
A tibble with feature indices and names included.
}
\description{
Modifies the limma top table to include feature indices and names.
}
