% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tutorial_and_template.R
\name{open_template}
\alias{open_template}
\title{open_template()}
\usage{
open_template()
}
\value{
If successful, opens the `template.Rmd` file in RStudio for the user to
interact with.
If `rstudioapi` is not installed or available, or the template file is
not found,
an error is thrown with a corresponding message.
}
\description{
This function opens the `template.Rmd` file in RStudio for
interactive use. The template file provides a structure for users
to quickly set up their personal analysis.
}
