% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tutorial_and_template.R
\name{open_tutorial}
\alias{open_tutorial}
\title{Interactive Tutorial for Getting Started}
\usage{
open_tutorial()
}
\description{
This function opens the `tutorial.Rmd` file in RStudio for
interactive use. Users can then run each code chunk step by step.
}
