% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{perform_clustering}
\alias{perform_clustering}
\title{Perform Clustering}
\usage{
perform_clustering(
  top_tables,
  adj_pthresholds,
  clusters,
  meta,
  condition,
  spline_params,
  mode
)
}
\arguments{
\item{top_tables}{A list of top tables from limma analysis.}

\item{adj_pthresholds}{A numeric vector of p-values.}

\item{clusters}{A list specifying clusters or "auto" for automatic estimation.}

\item{meta}{A dataframe containing metadata.}

\item{condition}{A character string specifying the condition.}

\item{spline_params}{A list of spline parameters for the analysis.}

\item{mode}{A character string specifying the mode ('isolated' or 'integrated').}
}
\value{
A list of clustering results for each level within the condition.
}
\description{
Performs clustering on top tables using specified p-values and clusters 
for each level within a condition.
}
\examples{
top_tables <- list(data.frame(feature_index = 1:10, adj.P.Val = runif(10)))
adj_pthresholds <- runif(10)
clusters <- list("auto")
meta <- data.frame(Time = seq(1, 10), condition = rep(c("A", "B"), each = 5))
condition <- "condition"
spline_params <- list(spline_type = c("n"), dof = list(3))
mode <- "isolated"
perform_clustering(top_tables, adj_pthresholds, clusters, meta, condition, 
                   spline_params, mode)

}
\seealso{
\code{\link{process_level_cluster}}
}
