% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_all_mean_splines}
\alias{plot_all_mean_splines}
\title{Plot All Mean Splines}
\usage{
plot_all_mean_splines(curve_values, plot_info, level)
}
\arguments{
\item{curve_values}{A dataframe containing curve values and cluster
assignments.}

\item{plot_info}{List containing the elements y_axis_label (string), 
time_unit (string), treatment_labels (character vector),
treatment_timepoints (integer vector). All can also be NA. 
This list is used to add this info to the spline plots. 
time_unit is used to label the x-axis, and treatment_labels
and -timepoints are used to create vertical dashed lines,
indicating the positions of the treatments (such as 
feeding, temperature shift, etc.).}

\item{level}{One of the unique values of the meta condition column. This is
a factor that separates the experiment.}
}
\value{
A ggplot object representing the average curves by cluster.
}
\description{
Generates a plot of average curves for each cluster, showing min-max
normalized intensities over time.
}
