% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_cluster_mean_splines}
\alias{plot_cluster_mean_splines}
\title{Plot Consensus Shapes}
\usage{
plot_cluster_mean_splines(curve_values, plot_info)
}
\arguments{
\item{curve_values}{A dataframe containing curve values and cluster
assignments.}

\item{time_unit_label}{A character string specifying the time unit label.}
}
\value{
A list containing a plot for every cluster
}
\description{
Generates composite plots of single and consensus shapes for each cluster
of curve values.
}
\seealso{
\code{\link{plot_single_and_mean_splines}}
}
