% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_consensus_shapes}
\alias{plot_consensus_shapes}
\title{Plot Consensus Shapes for Each Cluster}
\usage{
plot_consensus_shapes(curve_values, title)
}
\arguments{
\item{curve_values}{A dataframe containing the curve values and a cluster 
identifier.
                    The last column is assumed to be the cluster column.}

\item{title}{The base title for the plots, which will be modified by 
appending the cluster number.}
}
\value{
A list of ggplot objects, each representing a plot for a cluster.
}
\description{
This function takes a dataframe containing curve values and their associated 
clusters,
and generates a plot for each cluster using the 
plot_single_and_consensus_splines function.
Each plot represents both the individual curves and their consensus within 
the cluster.
}
