% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot Dendrogram}
\usage{
plot_dendrogram(hc, k)
}
\arguments{
\item{hc}{A hierarchical clustering object.}

\item{k}{An integer specifying the number of clusters.}
}
\value{
A ggplot object representing the dendrogram.
}
\description{
Generates a dendrogram plot for hierarchical clustering results, 
colored by clusters.
}
\examples{
\dontrun{
hc <- hclust(dist(matrix(runif(100), nrow = 10)))
k <- 3
plot_dendrogram(hc, k)}

}
\seealso{
\link[dendextend]{color_branches}, \link[dendextend]{as.ggdend}, 
\link[ggplot2]{ggplot2}
}
