% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(
  data,
  meta,
  condition,
  feature_names,
  all_levels_clustering,
  time_unit_label
)
}
\arguments{
\item{data}{A matrix of data values.}

\item{meta}{A dataframe containing metadata.}

\item{condition}{A character string specifying the condition.}

\item{feature_names}{A character vector of feature names.}

\item{all_levels_clustering}{A list containing clustering results for each 
level within the condition.}

\item{time_unit_label}{A character string specifying the time unit label.}
}
\value{
A list of ComplexHeatmap heatmap objects for each level.
}
\description{
Generates heatmaps for each level within a condition, showing z-scores of 
log2 intensity values, split by clusters.
}
\examples{
data <- matrix(runif(100), nrow = 10)
meta <- data.frame(Time = seq(1, 10), condition = rep(c("A", "B"), each = 5))
condition <- "condition"
feature_names <- paste0("feature", 1:10)
all_levels_clustering <- list(
  list(clustered_hits = data.frame(feature = 1:10, cluster = rep(1:2, 5)))
)
time_unit_label <- "[min]"
plot_heatmap(data, meta, condition, feature_names, all_levels_clustering, 
             time_unit_label)

}
\seealso{
\code{\link{ComplexHeatmap::Heatmap}}, \code{\link{dplyr::arrange}}
}
