% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_hyperparams_screen.R
\name{plot_limma_combos_results}
\alias{plot_limma_combos_results}
\title{Plot limma Combination Results}
\usage{
plot_limma_combos_results(
  top_tables_combos,
  datas,
  metas,
  spline_test_configs,
  meta_batch_column,
  time_unit = time_unit
)
}
\arguments{
\item{top_tables_combos}{A list of top tables for each combination.}

\item{datas}{A list of matrices.}

\item{metas}{A list of metadata corresponding to the data matrices.}

\item{spline_test_configs}{A configuration object for spline tests.}

\item{meta_batch_column}{A character string specifying the meta batch column.}
}
\value{
A list of results including hit comparison plots and composite 
spline plots for each pair of combinations.
}
\description{
Generates plots for pairwise comparisons of hyperparameter combinations 
using limma results.
}
\examples{
plot_limma_combos_results(top_tables_combos = list(table1, table2),
                          datas = list(matrix(1:4, 2, 2)), 
                          metas = list(data.frame(batch = 1:2)), 
                          spline_test_configs = 
                              list(spline_type = c("type1")), 
                          meta_batch_column = "batch")
                          
}
