% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_single_and_consensus_splines}
\alias{plot_single_and_consensus_splines}
\title{Plot Single and Consensus Splines}
\usage{
plot_single_and_consensus_splines(time_series_data, title)
}
\arguments{
\item{time_series_data}{A data frame of time series data.}

\item{title}{The title of the plot.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function takes a time series data frame, transforms it, and plots both
individual time series and their consensus using ggplot2.
}
