% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hits.R
\name{plot_single_and_mean_splines}
\alias{plot_single_and_mean_splines}
\title{Plot Single and Mean Splines}
\usage{
plot_single_and_mean_splines(time_series_data, title, plot_info)
}
\arguments{
\item{time_series_data}{A dataframe or matrix with time series data.}

\item{title}{A character string specifying the title of the plot.}

\item{plot_info}{List containing the elements y_axis_label (string), 
time_unit (string), treatment_labels (character vector),
treatment_timepoints (integer vector). All can also be NA. 
This list is used to add this info to the spline plots. 
time_unit is used to label the x-axis, and treatment_labels
and -timepoints are used to create vertical dashed lines,
indicating the positions of the treatments (such as 
feeding, temperature shift, etc.).}
}
\value{
A ggplot object representing the single and consensus shapes.
}
\description{
Generates a plot showing individual time series shapes and their consensus
(mean) shape.
}
\seealso{
\code{\link{ggplot2}}
}
