% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_rna_seq_data.R
\name{preprocess_rna_seq_data}
\alias{preprocess_rna_seq_data}
\title{preprocess_rna_seq_data()}
\usage{
preprocess_rna_seq_data(splineomics, normalize_func = NULL)
}
\arguments{
\item{splineomics}{An S3 object of class `SplineOmics` that must contain the 
following elements:
\itemize{
  \item \code{data}: A matrix of the omics dataset, with feature names 
  optionally as row headers (genes as rows, samples as columns).
  \item \code{meta}: A dataframe containing metadata corresponding to the 
  \code{data}. The dataframe must include a 'Time' column and a column 
  specified by the \code{condition}.
  \item \code{design}: A character string representing the design formula 
  for the limma analysis (e.g., \code{'~ 1 + Phase*X + Reactor'}).
  \item \code{spline_params}: A list of spline parameters used in the 
  analysis. This can include:
    \itemize{
      \item \code{spline_type}: A character string specifying the type of 
      spline. Must be either \code{'n'} for natural cubic splines or 
      \code{'b'} for B-splines.
      \item \code{dof}: An integer specifying the degrees of freedom. 
      Required for both natural cubic splines and B-splines.
      \item \code{degree}: An integer specifying the degree of the spline 
      (for B-splines only).
      \item \code{knots}: Positions of the internal knots (for B-splines).
      \item \code{bknots}: Boundary knots (for B-splines).
    }
  \item \code{dream_params}: A named list or \code{NULL}. When not 
  \code{NULL}, it can contain:
    \itemize{
      \item \code{dof}: An integer greater than 1, specifying the degrees 
      of freedom for the dream topTable.
      \item \code{KenwardRoger}: A boolean indicating whether to use the 
      Kenward-Roger method.
    }
}}

\item{normalize_func}{An optional normalization function. If provided, this
function will be used to normalize the `DGEList` object. If not provided,
TMM normalization (via `edgeR::calcNormFactors`) will be used by default.
Must take as
input the y of: y <- edgeR::DGEList(counts = raw_counts) and output the y
with the normalized counts.}
}
\value{
The updaed `splineomics` object, now containing the `voom` object, 
which includes the log2-counts per million (logCPM) matrix and 
observation-specific weights. Additionally, the splineparams are updated with
the identified optimal dof based on LOOCV, when dof = 0L (for auto-dof)
}
\description{
The `preprocess_rna_seq_data()` function performs essential preprocessing
steps for raw RNA-seq counts. This includes creating a `DGEList` object,
normalizing the counts using the default TMM (Trimmed Mean of M-values)
normalization via the `edgeR::calcNormFactors` function, and applying the
`voom` transformation from the `limma` package to obtain log-transformed
counts per million (logCPM) with associated precision weights. If you
require a different normalization method, you can supply your own
custom normalization function.
}
\seealso{
\code{\link[edgeR]{DGEList}}, \code{\link[edgeR]{calcNormFactors}}
}
