% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineomics_object.R
\name{print.SplineOmics}
\alias{print.SplineOmics}
\title{Print function for SplineOmics objects}
\usage{
\method{print}{SplineOmics}(x, ...)
}
\arguments{
\item{x}{A SplineOmics object created by the \code{create_splineomics} function.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The function does not return a value. It prints a summary of
the SplineOmics object.
}
\description{
This function provides a summary print of the SplineOmics object, showing
relevant information such as the number of features, samples, metadata,
RNA-seq data, annotation, and spline parameters.
}
\details{
This function is automatically called when a SplineOmics object is printed.
It provides a concise overview of the object's contents and attributes,
including the dimensions of the data, available metadata, and other relevant
information such as annotations and spline parameters.
}
