% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_limma_hyperparams.R
\name{process_combo}
\alias{process_combo}
\title{Process Combination}
\usage{
process_combo(
  data_index,
  design_index,
  spline_config_index,
  pthreshold,
  datas,
  metas,
  designs,
  modes,
  condition,
  spline_test_configs,
  feature_names,
  padjust_method,
  ...
)
}
\arguments{
\item{data_index}{Index of the data in the datas list.}

\item{design_index}{Index of the design in the designs list.}

\item{spline_config_index}{Index of the spline configuration in the 
spline_test_configs list.}

\item{pthreshold}{The p-value threshold for significance.}

\item{datas}{A list of data matrices}

\item{metas}{A list of metadata corresponding to the data matrices.}

\item{designs}{A list of design matrices.}

\item{modes}{A character vector containing 'isolated' or 'integrated'.}

\item{condition}{A single character string specifying the condition.}

\item{spline_test_configs}{A configuration object for spline tests.}

\item{feature_names}{A character vector of feature names.}

\item{padjust_method}{A single character string specifying the p-adjustment 
method.}

\item{...}{Additional arguments.}
}
\value{
A list of top tables from the LIMMA spline analysis.
}
\description{
Processes a single combination of data, design, spline configuration, and 
p-threshold to generate LIMMA spline results.
}
\seealso{
\code{\link{create_spline_params}}, 
\code{\link{run_limma_splines}}
}
