% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_generation.R
\name{process_plots}
\alias{process_plots}
\title{Process Plots}
\usage{
process_plots(plots_element, element_name, plots_size, html_content)
}
\arguments{
\item{plots_element}{A list of plots to be processed.}

\item{element_name}{A character string specifying the name of the element.}

\item{plots_size}{A list of sizes for the plots.}

\item{html_content}{The current state of the HTML content.}
}
\value{
Updated HTML content with the plots included.
}
\description{
Converts plots to base64 and appends them to the HTML content.
}
