% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{process_top_table}
\alias{process_top_table}
\title{Process Top Table}
\usage{
process_top_table(process_within_level_result, feature_names)
}
\arguments{
\item{process_within_level_result}{List of lists containing the limma 
topTable, fit, and optionally the voom
object. All of this is from one specific
level.}

\item{feature_names}{A non-empty character vector of feature names.}
}
\value{
A dataframe containing the processed top table with added intercepts.
}
\description{
Processes the top table from a LIMMA analysis, adding feature names and 
intercepts.
}
\seealso{
\link{modify_limma_top_table}, \link[limma]{lmFit}
}
