% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{process_within_level}
\alias{process_within_level}
\title{Process Within Level}
\usage{
process_within_level(
  data,
  preprocess_rna_seq,
  normalization_fun,
  meta,
  design,
  spline_params,
  level_index,
  padjust_method
)
}
\arguments{
\item{data}{A matrix of data values.}

\item{preprocess_rna_seq}{Boolean specifying whether to preprocess RNA seq}

\item{normalization_fun}{Function for normalizing RNA-seq raw counts.}

\item{meta}{A dataframe containing metadata, including a 'Time' column.}

\item{design}{A design formula or matrix for the limma analysis.}

\item{spline_params}{A list of spline parameters for the analysis.}

\item{level_index}{The index of the level within the factor.}

\item{padjust_method}{A character string specifying the p-adjustment method.}
}
\value{
A list containing the top table and the fit object from the limma 
analysis.
}
\description{
Performs a within-level analysis using limma to generate top tables and fit 
objects based on the specified spline parameters. Performs the limma spline 
analysis for a selected level of a factor
}
\seealso{
\link[splines]{bs}, \link[splines]{ns}, \link[limma]{lmFit}, 
\link[limma]{eBayes}, \link[limma]{topTable}
}
