% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{run_limma_splines}
\alias{run_limma_splines}
\title{run_limma_splines.R contains the exported package function run_limma_splines  
and all the functions that make up the functionality of run_limma_splines.
run_limma_splines performs a limma analysis, using splines, to assign a 
p-value to every feature of a time series omics dataset, to find out which
features are significantly changed over the time course.
Run Limma Analysis with Spline Interpolation for Hyperparameter Screening}
\usage{
run_limma_splines(
  data,
  meta,
  design,
  condition,
  spline_params = list(spline_type = c("n"), dof = c(2L)),
  padjust_method = "BH"
)
}
\arguments{
\item{data}{A dataframe containing the expression data, 
where rows represent features (e.g., genes) and columns represent 
samples.}

\item{meta}{A dataframe containing the metadata for the samples, 
must include a column for each factor in `factors`.}

\item{design}{A character string specifying the design formula for the Limma 
model.}

\item{condition}{The name of the factor column of meta. The factor divides
the experiment into different levels (unique elements of the factor column).}

\item{spline_params}{A named list, specifying the parameters for the splines,
such as type (ns or B-splines, dof, knots, etc).}

\item{padjust_method}{A character string specifying the method for adjusting 
p-values for multiple testing. Defaults to "BH" 
(Benjamini-Hochberg).}

\item{mode}{A character string specifying the analysis mode: 'isolated' for 
analyzing each level of a factor separately, or 'integrated' for comparing 
levels within a factor.}
}
\value{
A list containing three elements: 
        - `top_tables`: A list of top tables generated from within-level 
        analysis 
                        for each factor and level.
        - `ttslc_factor_only`: A list of top tables from between-level 
        comparisons 
                               for each factor, excluding time effects.
        - `ttslc_factor_time`: A list of top tables from between-level 
        comparisons 
                               for each factor, including time effects.
}
\description{
This function conducts differential expression analysis using the Limma 
package, 
incorporating spline interpolation to model the effect of various 
experimental 
factors across different levels. It supports both isolated and integrated 
modes 
for within-level analysis and between-level comparison, adjusting for 
multiple 
degrees of freedom corresponding to the factors under investigation.
}
\examples{
\dontrun{
  results <- run_limma_splines(data, meta, "~ 1 + Factor*X + Time", c(2L), 
                               c("Factor"), c("Gene1", "Gene2"), "isolated")
}

}
