% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineomics_object.R
\name{update_splineomics}
\alias{update_splineomics}
\title{Update a SplineOmics object with new values for stored variables or add new
variables}
\usage{
update_splineomics(splineomics, ...)
}
\arguments{
\item{splineomics}{A SplineOmics object to be updated.}

\item{...}{Named arguments with new values for fields to be updated or added.}
}
\value{
The updated SplineOmics object.
}
\description{
Updates a SplineOmics object by modifying existing fields or adding new ones.
}
