% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{within_level}
\alias{within_level}
\title{Within level analysis}
\usage{
within_level(
  level,
  level_index,
  spline_params,
  data,
  rna_seq_data,
  meta,
  design,
  dream_params,
  condition,
  feature_names,
  padjust_method,
  mode
)
}
\arguments{
\item{level}{The level within the condition to process.}

\item{level_index}{The index of the level within the condition.}

\item{spline_params}{A list of spline parameters for the analysis.}

\item{data}{A matrix of data values.}

\item{rna_seq_data}{An object containing the preprocessed RNA-seq data,
such as the output from `limma::voom` or a similar preprocessing pipeline.}

\item{meta}{A dataframe containing the metadata for data.}

\item{design}{A design formula or matrix for the limma analysis.}

\item{dream_params}{A named list or NULL. When not NULL, it must at least 
contain the named element 'random_effects', which must contain a string that
is a formula for the random effects of the mixed models by dream. 
Additionally, it can contain the named elements dof, which must be a int
bigger than 1, which is the degree of freedom for the dream topTable, and
the named element KenwardRoger, which must be a bool, specifying whether
to use that method or not.}

\item{condition}{A character string specifying the condition.}

\item{feature_names}{A non-empty character vector of feature names.}

\item{padjust_method}{A character string specifying the p-adjustment method.}

\item{mode}{A character string specifying the mode
('isolated' or 'integrated').}
}
\value{
A list containing the name of the results and the top table of
         results.
}
\description{
Processes a single level within a condition, performing limma analysis
and generating the top table of results.
}
\seealso{
\code{\link{within_level}}, \code{\link{process_top_table}}
}
