% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_splines.R
\name{within_level}
\alias{within_level}
\title{Performs the limma spline analysis for a selected level of a factor}
\usage{
within_level(data, meta, design, factor, level, DoF, padjust_method)
}
\description{
Performs the limma spline analysis for a selected level of a factor
}
